Improve kernel memory used calculation
CMA that is used by the kernel is allocated by kernel drivers, and
should naturally be accounted as part of kernel memory, so add it
as part of the getKernelUsedSizeKb() method.

GPU private memory is considered to be kernel memory, so add it
to getKernelUsedSizeKb() too.

While DMA-BUFs that aren't mapped in userspace and have no file
descriptors referring to them should also be accounted as part of kernel
memory, calculating that value is expensive and has been shown to
take up to 1 second at times.

Unfortunately, there are other callers of getKernelUsedSizeKb()
that don't need to account DMA-BUFs since they're not calculating
lost RAM, so to not degrade their performance, leave
the kernel DMA-BUF accounting where they are. They will be
consolidated in a subsequent patch.

Bug: 406327847
Flag: EXEMPT PURE_REFACTOR
Change-Id: I5c68e459ff25b2eee5c517087965ba902c1a4de3