Separate-out codepath that creates WAKE transition
The current sleep logic mixes multiple code flows together
into a mega function which, conditionally, creates a WAKE
transition.

This means we can't accurately distinguish the cause(s) of
this transition among all the code-paths that enter this
method.

This CL starts to narrow-down the conditions leading to this
by separating the "can make transition" part of applySleepTokens
from the "only going to sleep" part. This will then enable
walking up the call hierarchy to determine which call hierarchies
are already part of a dedicated action vs which one(s) are
specifically for making the WAKE transition.

Bug: 423701190
Test: Refactor, so existing tests
Flag: EXEMPT PURE_REFACTOR
Change-Id: I40b912ba0f321dfae0b67c193612f4f56c335b3d