AudioManager: add option to isolate focus for a uid
Addition of AudioManager methods to enter/exit audio focus
isolation for a given UID, with support for the exit mode.

When a uid's audio focus is isolated, its audio focus requests are
always granted, and other applications' requests are not affected by the
audio focus state of isolated uids. This is useful, for example, when
re-routing an applications audio to a different device. Enabling other
apps on this device to play simultaneously.

The isolation exit mode defines the behavior of a uid coming out of
isolation. In this version, there are two modes. In one of them, the
uid abandons any active focus requests and, in the other, the
uid attempts to retain an active request, if any.

Bug: 428808570
Flag: android.media.audio.audio_focus_isolation
Test: atest frameworks/base/services/tests/servicestests/src/com/android/server/audio/AudioManagerFocusIsolationTest.java
Change-Id: If777de6c7b196349e00d2c898c9c544a692cd5ce