Write settings for current user in VDM for ViewConfigurationParams
This change ensures that device-aware ViewConfiguration doesn't break
after the CoreSettingsObserver refectoring in ag/34714341. This is an
initial step to keep supporting device-aware ViewConfiguration for the
current user, the next step would be to support the same for all users
when a virtual device owner applies ViewConfigurationParams for a
virtual device.

Test: atest ViewConfigurationControllerTest
Test: atest VirtualDeviceViewConfigurationTest
Bug: 370720522
Flag: android.companion.virtualdevice.flags.viewconfiguration_apis
Change-Id: I3fe6eddf673ac3905e6b4862ff1284485e31ee07