Tie unified profile password atomically
The process of tying a profile's password to the parent
user's credential is not atomic and can result in
inconsistent state when the process is interrupted in
the middle.

This is because the new profile LSKF protector is committed
with a new random password before the tying operation. If
the tying operation is interrupted, it becomes impossible
to unlock the profile storage. Either the decrypt key is
not suitable for the password ciphertext, or it is but
the decrypted password is the old one and no longer unlocks
the new profile LSKF protector.

To avoid this inconsistent state, reorder the tying
operation to set up all unified profile password state
before committing the new protectorId. Assigning the new
profile protectorId now serves as an atomic commit of the
unified profile password credential.

Details:

* Password ciphertext is stored as spblob/$protectorId.pwd
* Keystore decrypt key alias now has .$protectorId suffix
* Migration is performed inline on retrieval of password

Bug: 412331826
Flag: android.security.enable_atomic_child_profile_lskf
Test: atest FrameworksServicesTests:com.android.server.locksettings
Test: Manual: flag off -> on, verify new spblob file
Test: Manual: flag on, change main user LSKF, reboot. Verified unlock
Test: Manual: flag on -> off, reboot. Verified unlock
Test: Manual: flag off, change main user LSKF, reboot. Verified unlock
Test: Manual: flag off, "Use one lock" off. Verified no pwd files
Test: Manual: flag off, "Use one lock" on. Verified legacy pwd file
Change-Id: Ie311b92d89da0a5e710c9fd298b706db52c6d3aa