Add support for the LHDC codec in BluetoothCodecConfig
This is intended to be a temporary patch while
a more definitive solution is being developed.

This change hard-codes the value of SOURCE_CODEC_TYPE_LHDC
in the audio framework, which will not be added as API.
A less temporary patch will involve using the public
CODEC_ID_LHDC value instead of the codec types, and a longer
term patch will involve exposing a new API for Bluetooth to
specify the Audio format.

Bug: 405925310
Test: TreeHugger
Test: Manual audio streaming test
Flag: com.android.bluetooth.flags.lhdc_codec_support
Change-Id: I83c632d29f27aeb6a6f505120f9d3387587bde99