Block signals and handle ANR
SIGUSR1 is used to inform ANR to the app, and SIGQUIT is used to
actually terminate it as a result of the ANR. This commit blocks SIGUSR1
and SIGQUIT (the default behavior is to terminate the process), and
spawns a thread to catch and report the signals. This design reflects
what SignalCatcher does. In non-native apps, ANR triggers GC to run on
the ART runtime and dumps some information, but we just emit some logs
for now. SIGPIPE is also blocked as in the Zygote implementation.

Test: m zygote_all
Test: Ensured SIGUSR1 and SIGQUIT are handled
Bug: 431634411
Flag: android.os.native_framework_prototype
Change-Id: I723acc45790ec799a25cdf2d6904846dbdbceb6f