Add framework-network-security-config Java sdk library
In this CL we are moving the NSPolicy class into a separate java sdk
library. This library is still part of the non-updatable frameworks code
but lives in framework/base/packages. The API provided by the NSPolicy
is also moved from core/api/current.txt to
packages/NetworkSecurityConfig/api/current.txt.

In a follow-up CL I will add a wrapper NSPolicy class to this library
that will allow us to redirect calls to the NSPolicy implementation in
Mainline. We will be able to switch between the current non-updatable
implementation and the one in Mainline using the
release_conscrypt_nsc build flag.

***NOTE***

Only the NSPolicy class is added to the java sdk library, not the
NSConfig classes inside platform/src/android/security/net/config.

The reason for this difference is that the NSConfig classes currently do
not expose APIs. Adding them to a separate java_sdk_library would
require us to mark and flag some of the classes and methods in the
package as System APIs. However, existing clients of those methods in
the platform would be broken until the flags are completely ramped up.

The wrapper NSConfig classes to redirect calls to the Mainline
implementation will therefore be kept inside the current
framework-network-security-config-sources filegroup, not in the java
sdk library.

API-Coverage-Bug: 443005615
Bug: 397646538
Test: m && vendor/google/tools/flashall
Flag: build.release_conscrypt_nsc
Change-Id: I302adb41618f4a1e88e42a05e054dab77968e737