[AOA FFS Implementation] Verify FFS kernel support for restorecon
This change adds a check which validates the kernel's support for
SELinux labeling of functionfs files, or if SELinux is not in enforcing
mode.

If SELinux is enforcing, we need to have ep1 and ep2 labeled as
u:object_r:usbaccessory_device:s0.

If the restorecon is not successful, either a permission issue exists,
or the kernel change to support individual file labels is not present,
and therfore we can not enable userspace AOA.

Bug: 407985367
Bug: 424156976
Bug: 416223903
Flag: android.hardware.usb.flags.enable_aoa_userspace_implementation
Test: manually validate the result with and without the associated
kernel change.
Change-Id: Ibdb42cd4d2f2a418162c24585c66fbccf85f4103
Signed-off-by: Neill Kapron <nkapron@google.com>