Add test for launching presentation with unfocused host task
Adds a test case to verify that a presentation can be successfully launched on a secondary display even when its host task on the primary display is visible but not focused.

This complements the existing test, which only covers the failure case where the presentation and the unfocused host task are on the same display.

Test: ATP tests passed http://go/forrest-run/L84100030017567708
Bug: 431235865
Flag: TEST_ONLY



Change-Id: I25980ad3e7adeaa804afb7d14d0750542b221e52