[1/n] Add topCompatActivityLeash to TransitionInfo.Change
The topCompatActivityLeash class will be added to a ChangeInfo in case of
letterbox activities in Task participating in a Transition.

topCompatActivityLeash contains the reference to the leash of the top
actiity, if present, in case of a transition for a Task. It'll be
present only if the top activity is letterboxed. The activity leash
will be used by Shell as parent surface for the letterbox surfaces.

Flag: com.android.window.flags.app_compat_refactoring_use_activity_leash_for_letterboxing
Bug: 443245627
Test: atest FrameworksCoreTests:TransitionInfoTest

Change-Id: I6e1a86d41b860d835fb11ee48e4708cc207f773c