Add test for updating split layout config on non-default display
This commit adds a unit test to verify that when `startIntent` is called for a non-default display, the `SplitLayout` configuration is correctly updated.

The test ensures that `updateSplitLayoutConfig` is effectively called by checking for the subsequent call to `mSplitLayout.updateConfiguration` with the appropriate display ID and configuration. This confirms the intended behavior when launching into split-screen on an external or secondary display.

Test: atest libs/WindowManager/Shell/tests/unittest/src/com/android/wm/shell/splitscreen/StageCoordinatorTests.java - http://ab/I66300010443041283 (With ag/35442980 reverted, failing) http://ab/I99400010443653502 (with ag/35442980 in the codebase, passing), AND ATP tests passed http://go/forrest-run/L19000030017625338
Bug: 431235865
Flag: TEST_ONLY

Change-Id: I9b6932643a1c68c52f2c80f94713f6c75bd8df1f