psc: Move LRU sequence logic to base class
This CL refactors the management of the LRU sequence number (`mLruSeq`)
by moving it from `ProcessList` to a new abstract base class,
`ProcessListInternal`.

Previously, `ProcessListInternal` was an interface. It has been
converted to an abstract class to hold the state (`mLruSeq`) and
related methods (`getLruSeqLOSP`, `incrementLruSeq`). `ProcessList`
now extends this class and calls the superclass methods to manage the
LRU sequence.

Bug: 442286304
Test: m services.core
Test: atest MockingOomAdjusterTests OomAdjusterTests FgsLoggerTest
Test: atest FrameworksServicesTestsRavenwood_ProcessStateController
Flag: EXEMPT PURE_REFACTOR

Change-Id: I7d1db3e7023f331d01108cdf8940031b157b2aa1