Remove dead code behind a flag
The code is behind a flag that has been ramped down for over a week. And
so far we have not seen negative impact. The logic or code behind this
flag is for old aconfig flag storage mechanism which is no longer used
in new aconfig storage mechanism. And it is in conflict with mainline
beta aconfig flags.

Bug: b/442853137
Test: m and atest
Flag: com.android.providers.settings.stage_all_aconfig_flags
Change-Id: I1cf56d567c52c9613192a26ee0edcf05ccdd386b