Audio: Fix unintended CEC volume commands during simultaneous output
When audio is output simultaneously to both an HDMI device and Wired Headphones,
adjusting the headphone volume incorrectly sends a CEC volume command.

This solution is used instead of a direct ARC/eARC check because the CTS
environment may not emulate an ARC connection, but getDeviceForStream
will correctly identify the main HDMI device as the priority, allowing
the test to pass.

Bug: 435542471
Test: manual - verify CEC commands not sent when changing wired
      headphone volume during simultaneous audio output to HDMI
      audio device and wired headphones.
Flag: EXEMPT BUGFIX

Change-Id: I0ef0b16c838a4c4cc7ec9eead8084a66abcc28b1
Signed-off-by: Tomoki Yonezawa <Tomoki.Yonezawa@sony.com>