Delay enforcing Telephony features with exceptions until Android 18
As of Android 16, the behavior of the Telephony APIs is not
consistent, and the UnsupportedOperationExceptions triggered when the
Telephony features are enforced may or may not be thrown, depending on
how the Telephony stack is configured.

On systems where these exceptions are thrown, both 0P/1P services and
3P apps may crash because they are not correctly catching and handling
them.

In order to mitigate all the possible 3P app crashes due to these
exceptions, the enforcement of the features is now delayed until
devices released with Android 18, while a solution is devised for the
inconsistent APIs.

This change also fixes how the vendor API level value is being
compared, given that its format is YYYYMM since Android 14-QPR3. The
previous check using the VANILLA_ICE_CREAM (SDK level) value should
have instead used 202404, as that was the vendor API level in Android
15. This check only worked because the format change to YYYYMM happened
right in that very same release.

Bug: 434622492
Test: atest
Flag: EXEMPT BUGFIX

Change-Id: I239a3a7205d44820ec10cb53d30cbba9f48eebcc