Extend autofill to support multiple concurrent users
AutoFillUI is a singleton that maintains a single instance of the
UI components (mFillUi, mSaveUi, etc.). When a new session's autofill
UI is triggered, it overwrites and dismisses the UI of the previous
session. Thus it doesn't support multiple concurrent session.

To fix this, this CL modifies AutofillManagerService to create and
manage a map that stores different AutofillUI instances, keyed by
user ID.
Each user will have its own AutofillUI instance.
When a Session is initialized, AutofillManagerServiceImpl retrieves
(or creates) the correct AutofillUI instance from the map, based on
the user ID.
The AutofillUI instance is responsible for managing its own UI
elements and callbacks, and displaying the UI on the correct display.

Bug: 280116881
Test: atest CtsAutoFillServiceTestCases
Test: atest --user-type secondary_user_on_secondary_display CtsAutoFillServiceTestCases
Flag: android.service.autofill.support_multi_user_multi_display

Change-Id: I60079d19283e8d2bd50df9117c747433bf2b12ee