Deprecate EnforcedAdmin in favor of EnforcingAdmin
The recommended way to represent an admin putting restrictions is using
EnforcingAdmin which is consistent with the implementation in
DevicePolicyManager. Deprecate the EnforcedAdmin class in
RestrictedLockUtils in favor of EnforcingAdmin. Mark all restricted
settings methods that take EnforcedAdmin as deprecated and add
documentation to recommend using EnforcingAdmin.

Flag: android.app.admin.flags.policy_transparency_refactor_enabled
Bug: 414733570
Test: m

Change-Id: I1ccd9d9d4152f7747244cf51d4b3ab3e1004ecf4