Fix the resizing state getting out of sync on some tiles
Changing a tile's anchors will snap its current value to the closest anchor. This normally only happens once, but some tiles will have a 1 pixel difference in anchor size between their small and large format, usually for tiles in the first column. This is enough to push the resizing state to the wrong size and block future resizing taps.

Part of the fix is to detect when it is safe to update the anchors, which would be when the user is NOT dragging. AnchoredDraggableState does not expose this state, so it has to be collected from a custom modifier that intercepts drag events without consuming them.

Flag: EXEMPT BUGFIX
Fixes: 444388840
Test: manually - resizing via drag/taps/undo
Test: ResizeTiles
Test: ResizingStateTest
Test: ResizingTest
Change-Id: Ib28b39b5184d696cdd6660f5f629e47128f5f4e0