Move setting sensitive notification state onto bubble data
Previously this was set on BubbleStackView which works well, however,
we don't have a nice way for Bubble Bar in launcher to know about this
unless we pass it through on the proxy -- which seems a bit much.

Instead, save this state on BubbleData, and when bubbles come in,
factor this on the suppressFlyout bit.

Flag: EXEMPT BUGFIX
Bug: 444398581
Test: atest BubbleDataTest
Change-Id: I5d7475411bae8ec51ed8ac9d874d619aa2cafd44