Fix a bug in IntentFirewall.joinPackages
joinPackages returns null in most common cases - short package
names that do not exceed the sufficient length (125). In such cases,
the code will not fall into the else if branch inside the for loop. So
it will exit out of the for loop with sb not empty. This leads it
to execute the return null statement at the end of the method.

Bug: 444971007
Test: manual - check the log
Flag: EXEMPT BUGFIX
Change-Id: I25cf3265718ca8b84c953a03ae3aefaf84defdd7