Don't skip onMeasureAction when initially expanding
Some outside condition has changed to cause this to set the width
to zero on start, which makes the date disappear until QS is
first expanded. Adding an additional check to the code that skips
the measure action when QS is not expanded.

Fixes: 443445858
Test: manually reproduced and verified the fix with the same steps
Flag: NONE small low risk bug fix
Change-Id: I176a17431ab0e707f85d8dced730e8d26bfc8325