[flexiglass] SceneContainerPlugin improvements
This class wasn't tested and now it is. This CL adds a comprehensive
test with 20 different scene and overlay combination. The approach taken
to list the expected flag values was to get to that scene/overlay
combination in pre-flexi and then run the command below to see which
flags are set (for dual shade overlays that don't have equivalents in
legacy, I just went to the equivalent of that overlay in single shade).

I then iterated on the production code until all test cases passed.

$ adb shell dumpsys activity service com.android.systemui/.SystemUIService | grep SysUiState -A 10

Bug: 439752998
Test: new test
Test: manually verified that flexiglass still looks and behaves normal
while locking, unlocking, and visiting all shade scenes and overlays
Flag: com.android.systemui.scene_container
Change-Id: I230ca3a9a42e29f04cd735a6faaac8d9650522d6

Change-Id: Ida90ef88b613a68f557d923fe8674c18f8e4bb01