Expose DevicePolicyManager.isDeviceManaged to other framework services
In particular, this will be used by the UserManager to allow deleting
the last full admin on managed devices. We expose this method via
DevicePolicyCache to ensure that there are no deadlocks when
DevicePolicyManagerService calls UserManagerService, e.g. to remove a
user, which in turn calls DevicePolicyManagerService to check if the
device is managed.

Flag: android.app.admin.flags.multi_user_management_device_provisioning
Test: m services.devicepolicy
Bug: 390162247
Bug: 394970894
API-Coverage-Bug: 442780973

Change-Id: I9c548653f1143e110d1cab713fa4e6ace6fd0e82