NPMS: fix crash due to NPE
On cellular-less devices where the 'config_force_phone_globals_creation'
flag is set, the Telephony stack is partially alive even if there is
no real baseband. In this scenario, CarrierConfigManager will be
unavailable.

This change checks the CarrierConfigManager is not null before calling
its methods, to avoid crashing the process. The same approach was used
in the phone process itself due to the same reasons.

Bug: 445542097
Test: atest NetworkPolicyManagerServiceTest
Flag: EXEMPT BUGFIX

Change-Id: I74ef50cae6bf73005361dd9b3680c2f45ca1939d