Validate resourceType
Make sure that only valid values are set for resourceType by using an
enum in the proto. This also makes using the proto simpler since the
values are well known.

An enum entry for 0 is required by proto, but will not be used.

IFTTT will help us to keep the Android API and the Proto enum in sync.

NO_IFTTT=No change to existing scopes

Bug: 445330343
Test: atest frameworks/base/tools/processors/devicepolicy/test/src/android/processor/devicepolicy/test/PolicyProcessorTest.kt
Flag: android.app.admin.flags.policy_streamlining
Change-Id: I965602bfe268566b2fbd634f38000135fa5ee3d5