Make DisableFlagsInteractor and DisableFlagsRepository per display
They are now @PerDisplaySingleton, and available in the
SystemUIDisplaySubcomponent. When used "inside" of the Subcomponent,
they can be injected directly with the @DisplayAware qualifier.

When accessing from "outside", they can be accessed via the
SystemUIDisplaySubcomponent property.

To access them from the shade code, DisabledFlags is exposed via
ShadeStatusBarComponentsInteractor, using a Flow, so that it always
contains the latest, based on which display the shade currently is on.

Test: atest SystemUITests
Test: Build & Run SystemUI and add and remove displays
Bug: 435653753
Fixes: 437310166
Flag: com.android.systemui.disable_flags_per_display
Change-Id: I95d3be81a60a6af48849ae8654c19376b5fd25f3