[flexiglass] Show/hide nav bar based on flexi state
Before this change, nav bar showing or hiding was controlled in a
distributed fashion and the logic related to whether to show it for the
keyguard (lock screen, bouncer, and occluded scenes/overlay) was done in
StatusBarKeyguardViewManager.

This CL disables the setting of the nav bar visibility in the legacy
manager class and adds a centralized side effect to the SceneContainer
composable that should work for any scene or overlay. The logic simply
checks if back navigation is available in the topmost scene or overlay
and sets the nav bar visibility accordingly.

Fix: 439752998
Test: manually verified by turning on 3-button navigation mode in the
Settings app and navigating between all the different flexiglass scenes
and overlays
Flag: com.android.systemui.scene_container

Change-Id: I63b2173066427aba72e6d0bdf1b8c11bc2664dab