Refactored HsuDeviceProvisionerTest.
That class currently sets just one setting, and its unit test
is exercising the whole workflow (registering a callback and changing
the settings once the callback is triggered), which makes it harder
to add tests for new settings.

This CL refactors the unit test so each of these 2 steps is tested in
isolation.

Bug: 424203846
Test: atest FrameworksMockingServicesTests:HsuDeviceProvisionerTest
Flag: TEST_ONLY
Change-Id: Ie8087a2fc40838edd7da36131840476bd7495666