Override LocalContext with selected user's Context for password bouncer
The OutlinedSecureTextField reads the SHOW_PASSWORD system setting to
determine whether it should reveal the last character briefly while
typing. For that, it accesses system settings from the contentResolver
provided by the LocalContext.current.

SysUI runs in user 0, but the user has changed their setting via
settings app for their own user (e.g. user 10, 11, 12). Therefore, the
OutlinedSecureTextField would always read the setting for user 0 and
ignore the user's choice.

To fix this, we introduce a local override of LocalContext for the
selected user, very similar to the prior art done with
SelectedUserAwareInputConnection.

SelectedUserAwareInputConnection can not be replaced with
SelectedUserAwareLocalContext because InputMethodManager reads the
default user ID from UserHandle.myUserId(), which is determined by the
current process and not the LocalContext.

Fix: 445981272
Flag: com.android.systemui.scene_container
Test: manual - changing the show_password setting for user 0 and 10 via adb and verifying whether last character is revealed while typing
Change-Id: I76c8c70b86d237e338740390ba8c176cbb78317e