Refactor atrace database tracing to use per-connection async track
The previous system uses the Trace.asyncTrace* APIs for each SQL
operation. This results in having async tracks per query instead of
having the temporal sequence of queries executed in a single track.
Refactored the tracing logic in SQLiteConnection such that async tracks
are created per SQLiteConnection objects. Hence, all SQL operations
executed on a connection will be visualized in the same track,
regardless of the operation executed.

The serialization of events is guaranteed because the connection objects
are always attached to a SQLiteSession which in itself is attached to a
singular thread. This means that any time a SQLiteConnection object is
executing, it is owned by a single thread.

One thing to note is that even though the serialized execution happens
entirely within one thread, we cannot use the thread track because the
execution of a SQL transaction could span more than one function. Meaning
it is possible that the child trace slice of the transaction outlives the
parent slices of individual functions.

Also, instant events were added to the calling thread tracks, in order
for users to manually map which thread executed which query in a
SQLiteConnection async track. The instant events are only added for
transactions or SQL queries executed outside of transactions. Queries
within transactions don't have corresponding instant events.

Apart from this SQL transaction slices were fixed, since they weren't
being generated given the lack of end slices and improper check of the
operation.mTraced field.

Flag: EXEMPT BUGFIX
Test: capture a perfetto trace with 'atrace_categories: database'
Bug: 445773416
Change-Id: Ibd6507bf86e3d61dc9bd8b3817c378f9a55476b5