Extract userId, package & sys entity into AdminKey
This way static information identifying admin entity is stored in
AdminKey while all other attributes are left in EnforcingAdmin.

ComponentName is passed into AdminKey for legacy DA pacakges, but it is
not moved there completely: currently POs and DOs also have component
name set in EnforcingAdmin, but we don't want that to be part of the
key, because ultimately DO package and the same package as a role holder
shoudl be treated as the same admin. So component is not set for
AdminKeys representing DOs and POs.

In case of system entity, mPackageName in EnforcingAdmin was set to an
empty string. In AdminKey it is set to null, but the getter returns an
empty string in those cases to maintain the same behavior.

Bug: 336315393
Test: TH
Flag: EXEMPT PURE_REFACTOR
Change-Id: Icf59f954d44a10d33b908c37ccde67ae95161e35