[flexiglass] Show/hide nav bar based on flexi state
Note: this is a second version of ag/35830961 which was reverted due to a failing unit test when flexiglass is on. The only difference between this version and the original is the @DisableSceneContainer annotation that was added to the test file.
Change-Id: I5991ca42c833156186326565e3b50954dc1a2090

---

Before this change, nav bar showing or hiding was controlled in a
distributed fashion and the logic related to whether to show it for the
keyguard (lock screen, bouncer, and occluded scenes/overlay) was done in
StatusBarKeyguardViewManager.

This CL disables the setting of the nav bar visibility in the legacy
manager class and adds a centralized side effect to the SceneContainer
composable that should work for any scene or overlay. The logic simply
checks if back navigation is available in the topmost scene or overlay
and sets the nav bar visibility accordingly.

Fix: 439752998
Test: manually verified by turning on 3-button navigation mode in the
Settings app and navigating between all the different flexiglass scenes
and overlays
Flag: com.android.systemui.scene_container

Bug: 446075156
Change-Id: Ic88fdb577abc044e70d1f5af79021183f1663e25