[flexiglass] Fixes dual shade edu tooltip position
The problem was that the popup was being pushed down to account for the
display cutout window inset - not necessary in dual shade.

Also swaps the order of the offset and onGloballyPositioned modifiers in
the code that reports the position of the status bar chip to reduce the
number of updates handled by the system (as the offset is a looping
bounce).

Fix: 446023095
Test: manually verified that the vertical gap between the popup and the
status bar chips is gone
Flag: com.android.systemui.scene_container

Change-Id: I6cc218fee6db55925c0c626b9ad6860d844a9825