[PiP on Desktop] Let PipTransition handle OPEN transitions
so that it can handle a Content PiP request, which comes in as an OPEN
transition. Previously, we added bypassing letting PipTransition handle
an OPEN transition while in Desktop mode in
If34483f4a2eaf2af5f9c9a05bed5867455187799 to prevent auto-enter PiP when
a new task is launched.

However, with the current implementation with Desks, undefined launches
should be assigned to a desk root and have its windowing mode be
resolved to freeform in Core, so this workaround is no longer needed.

Bug: 440427561
Test: Manual - launch YT in DW and play a video. Launch another new task
via the Taskbar and verify YT does not trigger PiP.
Test: Manual - using ApiDemos apk, trigger Content PiP. Verify PiP is
entered as intended.
Flag: EXEMPT BUGFIX

Change-Id: Ibade16d51d6e74038d26b6e4e1da431e8f10d757