Large screen screenshots: Extract PreCaptureToolbarViewModel
The PreCaptureViewModel is growing large and contains state for the
entirety of the pre-capture UI. Extracting a new child view model for
the toolbar helps organize the state and separate concerns.

Additionally, removes the toolbar button item generation out of the view
model and into the Composable to simplify the updates via depenendecies.

Bug: 412421893
Test: atest PreCaptureViewModelTest
Test: atest PreCaptureToolbarViewModelTest
Flag: com.android.systemui.large_screen_screencapture
Change-Id: I17d7d07fb2307522f853e096ceaf21963be0233a