Add Advanced Protection config for feature availability
The new config allows OEMs to override feature availability. The config
is defined via XSD schema. A new class AdvancedProtectionConfigLoader
uses the generated classes and parser from the XSD compiler. The config
is loaded in AdvancedProtectionService to decide whether hooks /
providers need to be added and if they're available in the
getAdvancedProtectionFeatures API.

Bug: 438955082
Test: atest AdvancedProtectionManagerTest
Test: atest AdvancedProtectionConfigLoaderTest
Test: atest AdvancedProtectionServiceTest
Test: atest DisallowCellular2GTest
Test: atest DisallowInstallUnknownSourcesTest
Test: atest MemoryTaggingExtensionTest
Flag: android.security.aapm_api_v2
Change-Id: I5ab5c33ed3774cf4620185a9a15580ac4a0b4dd7