psc: Move UID storage implementation to base class
Convert `ActiveUidsInternal` from an interface to a concrete base class
that handles the storage and management of active UIDs.

This CL refactors `ActiveUids` to extend `ActiveUidsInternal`,
delegating the core storage logic to the superclass. `ActiveUids` now
focuses on its specific responsibility of notifying observers, which
simplifies the code and improves encapsulation.

Bug: 446555028
Test: m services.core
Test: atest MockingOomAdjusterTests OomAdjusterTests
Test: atest FrameworksServicesTestsRavenwood_ProcessStateController
Flag: EXEMPT PURE_REFACTOR

Change-Id: I55d35a90b77821210ddf366e88f1882d3be50661