Use a Map for algorithm lookup in VpnConnectivityMetrics
Refactor the algorithm mapping in VpnConnectivityMetrics to use a
Map instead of a SparseArray and swap the key and value. This is
because SparseArray's `indexOfValue()` uses reference equality
(==), which can fail to find an algorithm even if the string
value is the same but the instance is different. This change
provides a more correct way to look up the integer index for a
given algorithm string.

Bug: 429288223
Test: atest FrameworksVpnTests
Flag: EXEMPT BUGFIX
Change-Id: I03229ae9076b80b4148e142e4b66af4ea4523d30