Do not use InputChannel with Activity
ViewRootImpl instantiates an InputChannel instance and passes it to
WindowManager to get it initialized before passing it to InputEventReceiver,
but that's not implemented on Ravenwood, so apparently the change
I4bbc5653f9 broke the experimental Activity inplemtation.

Fortunately, there's a window flag to disable InputChannel, so
let's use it as a stop-gap solution.

Bug: 447182552
Flag: TEST_ONLY
Test: $ANDROID_BUILD_TOP/frameworks/base/ravenwood/scripts/run-ravenwood-tests.sh -r
Change-Id: Id85b836c7923515d23530508d0fec5ac7d7c26af