Remove legacy mConfigWillChange
It was added in I4e062093cbcfbc919307799a794846db2920216f to
avoid intermediate config change of the current top activity
(will be pausing) before updating global config. However, since
new configuration won't dispatch to an activity that is going
to be invisible, the mConfigWillChange becomes no-op. Besides,
most cases ensureActivityConfiguration won't be called in the
current scope of mConfigWillChange when launching in fullscreen
windowing mode because a the visibility/config update happens
after the pausing activity is done.

Bug: 163976519
Flag: EXEMPT PURE_REFACTOR
Test: WmTests
Change-Id: Ib39c276f51d3bed0083b6079240de5147e3891ab