Retrieve and interact with screencapture saved location from ViewModel
We want to be able to retrieve the custom save location found within
disk. This way, we are able to display the location in the UI and also
use it for the functionality (saving the screenshot) later on. We also
want to be able to change this location. This CL creates an interactor,
allowing us to easily interact with the data within the repository from
the viewModel. This CL also has a placeholder UI (will be changed in future CL) within CaptureSettingsMenu to allow us to test (view and modify) the location along with unit tests.

Bug: 444278555
Test: Manual, PreCaptureToolbarViewModelTest, LargeScreenCaptureParametersInteractorTest
Flag: com.android.systemui.large_screen_screenshot_save_location
Change-Id: Idf7b22c12dea64f3d66ea6aa93e7d2ac57b5c223