Wait for TaskView relayout before start jumpcut switch animation
1. When the opening Bubble infates, it sets TaskView's initial layout
   height to MATCH_PARENT, which results its boundsOnScreen to fill
   display.
2. The TaskView's layout will later be updated when applying BubbleData
   update.
3. When actual animation is played, the `prepareOpenAnimation` will
   send a WCT to set the Task bounds to match TaskView's boundsOnScreen.

Before, after the layoutParams is updated in step#2, the relayout and
animation start can be racing, which often result a WCT to incorrectly
update the Task bounds.

Now, we wait for the relayout from step#2 is done before start the
animation.

Bug: 417848405
Test: atest WMShellUnitTests:BubbleTransitionsTest
Flag: com.android.window.flags.fix_bubble_trampoline_animation
Change-Id: I6bd845614214140ac3ca990c04fffd50d5d24e2e