Pass updated recording parameters to the service
This CL moves ScreenCaptureRecordParametersRepository to the ScreenCaptureScope which allows it to persist throughout the recording. This makes parameters consistent with the recording service. Now to keep changes consistent we only need to pass updates to the service. And this is what ScreenRecordingServiceInteractor does now.

I've also updated ScreenRecordingParameters to be parcelable because it's a more scalable approach and allows to pass extra parameters to the service without changing the aidl interface.

Flag: com.android.systemui.new_screen_record_toolbar
Bug: 443929609
Test: manual on foldable: Start recording with show_touches enabled. Change it throughout the recording and observe changed behaviour
Change-Id: I4744bf8db697ae0032151149e0ce0a817f3d7526