Respect opt-outs from force-rotate in other camera compat treatments.
When apps opt-out from camera compat treatent via manifest property (or are opted out using per-app overrides) it is often for one of two reasons:
1. The app work well without the treatment (false positive),
2. The app has severe issues with the treatment applied - this is often crashing or severe UI issues on configuration change.

Force-rotate treatment is being deprecated, and replaced with a simulate-requested-orientation treatment, which is more versatile. Reasons 1. and 2. are not expected to change in the new treatment, therefore we want to avoid extra work from the apps and per-app-overrides owners to opt-out these apps again.

Flag: com.android.window.flags.camera_compat_unify_camera_policies
Bug: 432218134
Test: atest WmTests:AppCompatCameraOverridesTest
Change-Id: If0ee5d82e0b028a9c6455174ab3b9d66c02e93d8