Skip LaunchOrConvertToBubble when it is opening existing Bubble
Before, LaunchOrConvertToBubble is expected to launch a new Bubble for a
given Intent, but didn't consider the case where the Intent is resolved
to an existng Bubbled Task.

Now, when this happens, it will skip the LaunchOrConvertToBubble, and
let DefaultMixedHandler to find the next matched handler if any.

Bug: 448009266
Test: atest WMShellUnitTests:BubbleTransitionsTest
Test: atest WMShellUnitTests:DefaultMixedHandlerTest
Flag: com.android.window.flags.fix_bubble_trampoline_launch_twice
Change-Id: I2acd006e6c42bf48f60ed28b036d074368c141af