Disable LetterboxScrollProcessor in freeform.
LetterboxScrollProcessor should be enabled in freeform windowing, as reachability is not a concern here. Additionally, apps in freeform have extended touch region to the window header (to be able to render tabs), and these events should not be modified for scrolling on letterbox.

Flag: com.android.window.flags.scrolling_from_letterbox
Test: atest FrameworksCoreTests:LetterboxScrollProcessorTest
Fixes: 381099873
Fixes: 381707503
Change-Id: I0eedadc728789c80fadd7134cc65e43a662a08e1