Screen capture: Avoid closing large-screen UI on click
On small screen devices, the screen capture UI should behave like a
dialog in that it dismisses when clicked outside. However, for large
screen devices, this resulted in unexpectedly closing the UI when a
a non-actionable element is clicked. Prevent the UI from closing in this
way for large screen only.

Bug: 448418000
Test: Manual
Flag: com.android.systemui.large_screen_screencapture
Change-Id: Iafe18476b4054de9f1cf9f5d82bf68f368805d3c