Disable rest of the animations on UserSwitchingDialog on timeout.
There is no need to keep pushing for UserSwitchingDialog animations
if one of the animations timeout. Because this means the UI thread
is under load, and animations are not important in that case.

This also fixes CalledFromWrongThreadException when show animation
timeouts, and the progress (spinner) animation is triggered on the
timeout call which is on the background thread.

Bug: 447744870
Test: atest UserControllerTest
Flag: EXEMPT BUGFIX

Change-Id: Idd0dfcefc3c5a5b0e040003780c511ea9bfce9bd