BouncyBall: Use a metric template
We reduce duplicatation in our metric specs by using a template.

This forces us to rename the fields since we were using
"bouncyball" as a suffix instead of a prefix.

We take this opportunity to move away from "missed_frame" as
a name, since this is misleading.  For example, a test run
could have a single bad frame where it misses both the app and
SF CPU deadlines, but only drops one frame.  Our metric will
return "2" for this, though.  So we more properly label this
as a jank instance, and one which is a test failure.

This also gives us room to tally some non-failing (but
informative) jank instances (like those involving buffer
stuffing) in the future.

NO_IFTTT=Moving a line as part of a refactor, not changing its contents

Bug: 448176306
Test: Ran trace_processor_shell on multiple traces with this spec, including ones with non-zero failing jank instances
Flag: EXEMPT TEST_ONLY
Change-Id: Idc4fc16fabbacf86d0faf45a0a883d5f80a904e3