Override the XML defaults of device items on the details tile
In this CL, BluetoothDetailsContentManager replaces the device item
layout's default params with the values read from res.

In addition, this CL adds dimension resources for the Bluetooth tile
details device entry. These resources will be later overridden when
the desktop sizing feature is on.

Typically, we use dimen res directly in the layout XML. However,
the device item layout is shared in two places: (1) dialog (2) details
tile. We are only planning to change the specs of the details tile.
Therefore, this CL customizes the device item layout in the controller.

Bug: 438560671
Flag: com.android.systemui.qs_tile_detailed_view
Change-Id: I3c805c199b072cf4df835a8c87ce9ebfa256bd91