[Bubbles] Fix FloatingToBarConversion finish callback
FloatingToBarConversion didn't call the finishCallback, causing
the transition system to halt when it handles animations independently.
This only happens in rare cases such as quickly folding and unfolding,
because the transition is usually merged into the unfold transition,
but when the issue happens, it blocks all future transitions and
the system gets into a broken state.

Bug: 441341037
Test: Manual
Flag: EXEMPT BUGFIX
Change-Id: I19ab4b316447451e030449d89e9f476c3cb90d27