Address dream end and teardown logic execution.
DreamOverlayService#isDreamInPreviewMode requires a client to be set.
Otherwise an illegal state exception can occur. The client can be unset
in two scenarios: the DreamService is torn down and the dream has ended.
Both these scenarios execute on the main thread (destroy by default and
endDream by marshalling).

A problem arises if logic is spread across execution loops. The existing
destroy code cleared the current client reference and then ran the
internal reset logic on the executor. As a result, queued code on the
executor has a chance to run in between in this inconsistent state.

This code changes the execution so all involved destroy logic runs
together. Additionally, the lifecycle is now updated to consider dream
ended as a paused state. The state is now used before considering
whether to call #isDreamInPreviewMode.

Test: atest DreamOverlayServicetest
Fixes: 432039567
Flag: EXEMPT BUGFIX
Change-Id: I996b298fd6f456532b9f3c691b72833a02733f3b