Allow APEXes to request full-fledged App Info UI for APKs they contain
Before this change, a limited version of the App Info UI was shown
for all APKs-in-APEX, in which the AppInfo buttons (i.e., Force Stop,
Disable, etc.) were hidden, and the storage management buttons were
disabled.

While this limited App Info behavior is required for mainline APEXes,
it should not apply all APEXes (e.g., the behavior of the App Info UI
for GmsCore should not change, when the GmsCore APK is moved to the new
GMS APEX).

Allow an APEX to opt the APKs it contains out of the limited App Info
behavior, via the inclusion of a <meta-data> value in the APEX's
AndroidManifest.

To ensure consistent UI behavior across all mainline modules, a
server-side check will prevent this <meta-data> from being included in
any mainline APEX.

Bug: 449011063
Flag: com.android.settingslib.flags.enable_limited_app_info_metadata_for_apex
Test: atest SettingsRoboTests
Change-Id: Ib05f3717758a93209c212d59aa8a8cc070438cf2