setOemUnlockEnabled allows SYSTEM not just Admin
setOemUnlockEnabled checks that the caller is an Admin user.
Historically, the System user has always been considered an Admin user,
so this always included user 0. But on some devices, user 0 will not be
an Admin user; for historical reasons, we still apparantly need to allow
user 0 to call this method, even though it isn't an Admin. So we relax
the requirement slightly. Note that there is already a separate
requirement that the caller have OEM_UNLOCK_STATE, so this relaxation is
not large.

Bug: 437244533
Bug: 443134869
Test: atest android.platform.tests.ProfileIconTest#testToVerifyGuestProfile on Automotive
Flag: android.multiuser.hsu_not_admin
Change-Id: I4a017421c846390670aea5f95a75694eb586aa07