Do not set the avrcp volume when absolute volume is set
This can cause a race condition since the AvrcpVolumeManager will send
the initial volume on device connection with onSetBtActiveDevice but will not reapply it.
On the other hand when BT is setting the device volume behaviour with avrcpSupportsAbsoluteVolume we will also set the currently stored AVRCP volume which can override the one from onSetBtActiveDevice.

Test: manual
Bug: 448753996
Flag: EXEMPT BUGFIX
Change-Id: Ifa6ba76244042b5fe888556a1ee590e66dbecca6