Ensure transition mode is OPEN for launch a window not in the hierarchy.
A task can be removed from the hierarchy but remain in Recents. Another
similar case is when an app's process is killed but its activity record
remains in the system server. In either case, the system has to create
a new window when launching the task again, and since the window did
not exist in the hierarchy, the mode of that change should be OPEN
instead of TO_FRONT.

Bug: 452214913
Flag: com.android.window.flags.promote_existence_changed_state_to_parent
Test: verify when a task is removed from hierarchy, when launch the task
again, the mode of that change should be OPEN.

Change-Id: Ib6a031f76af618a39a62f8fb640b5ba80e48906f