audio: do not unmute on set volume for PC
In CL:33305462 we have changes to not unmute for PC when volume is
adjusted downwards (i.e., on the volume down key).

For USB headsets, we use a custom handler and need to use the set volume
API due to possibly needing to move more than 1 step in each volume
adjust event (e.g., when the device has less HW vol steps than the UI).

Currently, on set volume, it unmutes. In this CL, we force that unmute
not to happen, and the desktop volume handler should explicitly unmute
when needed.

Bug: 451885995
Test: Mute, then set volume, make sure it's still muted.
Flag: EXEMPT BUGFIX
Change-Id: Ia27d32e070dbdf81519a0f27f77783badea74869