Animate dismiss PiP below display bounds
When animating dismiss PiP, it was being animated
based on the movement bounds; however, this creates
a non-ideal experience when for instance IME is present.

In general, the intention of drag-to-dismiss PiP is to animate
it off the screen bounds, so make sure to use that instead.

Bug: 452235541
Flag: EXEMPT BUGFIX
Test: enter PiP, open IME, then drag to dismiss PiP
Change-Id: Ibcfd3a8dabe22d5a4969f86019981852a285caf4