[1/N] Implement remembered bounds prototype
This CL implements a prototype of remembered bounds which supports the
minimal functionality of the feature; 1) to record the bounds ratio on
bounds changes and 2) to restore the bounds on the next app launch.
We record the “bounds ratio” which is a ratio of the bounds relative to
the stable bounds, which makes the remembered bounds to be display
resolution-agnostic.

Flag: com.android.window.flags.enable_remembered_bounds
Bug: 452164820
Test: DesktopTasksControllerTest
Test: DesktopTasksTransitionObserverTest
Test: DesktopRepositoryTest
Change-Id: I3288dc49a1bae2ae9c2abcbdef316b17057a16d2