[PL] Support NORMAL_APP_LAYER request
In the previous CLs ag/36168627 and ag/36348516 AoT changes were supported
to remove the task from a desk. This CL brings the task back to the desk
as soon as NORMAL_APP_LAYER request is received. This CL skips **a lot**
of flows that are going to be implemented in follow-up CLs.

NormalAppLayerHandler uses DesktopTasksController to populate wct
changes and transitions related to the desktop windows. It doesn't
satisfy callback yet and that'll be done in a separate CL in scope of b/448091682.

PinnedLayerController unpin logic is extended to consider a case when
there's a request to swicth to another layer. NormalAppLayerHandler is
invoked here because there are only 2 layers for now and that's just
simpler to do, so again handling normal layer requests will be in
follow-up.

Bug: 448091682
Bug: 449681882
Bug: 399376001
Flag: com.android.window.flags.enable_interactive_picture_in_picture
Test: atest WMShellUnitTests:PinnedLayerControllerTests
Screencast: https://screencast.googleplex.com/cast/NDc0OTU0MjA0NDk5MTQ4OHwxYzlmZjJkZC03Mw
Change-Id: If323400d3dba3d067ad9938d30da6e0323cc8b74