[sb] Remove StatusBarRootModernization (1/?)
This is the biggest chunk of work to remove this flag

* Remove CollapsedStatusBarFragment
* Remove related classes (logger, fragment provider, dagger module)
* Status bar initialzation is now un-flagged and simpler
* Compose is always enabled in status bar
* Tests have been updated / removed as necessary

This might be the only CL needed. But we can now evaluate the usage of
FragmentHostManager and its related components.

Note that the aconfig flag is left in for now, to try to minimize the
number of affected repos. That CL will follow.

Test: atest SystemUITests
Bug: 444212994
Flag: EXEMPT FLAG_REMOVAL
Change-Id: I7a0be022c3462048cd16a2398242a3aa724bfd6d