Fix showing the IME with preventImeStartup
When config_preventImeStartupUnlessTextEditor is set, the IME process is
only started while an EditText is focused, and otherwise it is not
running (signalled via startInput calls from InputMethodManager).

In [1] a new startInput call was added to allow restarting the IME after
its package data is cleared, which does not pass any startInput flags.
As this is sent after the first startInput, it will override the flags
received on the system_server, incorrectly signaling that no EditText is
focused, and triggering the flow to destroy the IME process.

While there are multiple similar startInput calls with no flags, these
pass a null view parameter, leading to the startInput flags being
resolved before the call is sent to the system_server. As this new call
uses a non-null view, this flag resolving is skipped. This fixes the
issue by manually computing the arguments, similar to the call from
onPostWindowGainedFocus, and thus allowing the IME to start and show.

This uncovered another issue with the flow for destroying the IME tied
to this config, namely that it only unbinds the IME, but does not send
an unbind signal to the IME Client app. While this could lead to various
issues, an immediate effect was that the boolean flag checked by the new
startInput call (from [1]) was not reset. Because of this, the flow
would only be triggered the first time the IME would be requested
(blocking starting the IME process), and then be skipped on following
attempts (allowing starting the IME process), leading to more
inconsistencies.

This fixes the issue by sending an unbind signal to the IME client,
after the IME process is unbound due to no EditText being focused. Note
that as part of the startInput handling, this would send a setActive
signal to the IME client before the preventImeStartup logic was checked.
This setActive call would then issue a new startInput signal, which
together with the new unbindInput signal, would lead to a loop of
startInput calls. Due to this, the preventImeStartup logic is placed
before the setActive call, and before the new IME Client can be set.

Before this fix, only the first show request since boot would fail, as
up to that point there would be no ImeInsetsSourceProvider created,
slowing down the show request processing. Moreover, this requires an
exact timing between the first startInput signal reaching the
system_server, the first show request setting the requestedVisibleTypes,
but failing due to the provider not being created (as the IME was not
started yet), and a second show request being sent to trigger the
startInput call from [1], but before the IME process could be started.

  [1]:  If933ec22307478a78f1eb023f9eba960870dc8ea

Flag: EXEMPT BUGFIX
Bug: 451934553
Test: atest FocusHandlingTest#testDestroyImeAfterEditorLostFocusWithPreventImeStartup
  FocusHandlingTest#testDelayedAddEditorStartsInput
  FocusHandlingTest#testNoEditorNoStartInput
  manual: with the config enable, tap on the search icon to show the IME
  on the messaging app, as the first IME show request since boot
Change-Id: I4a276343a6292d3c4fd06dab4dc94c09cc61b99c