Remove `bal_check_unpinned_activities` flag.
This change hardcodes the behavior to check each Activity for its pinned state when determining if a process has a visible, not-pinned activity.
The flag definition and tests for the old behavior are removed.

Reasoning behind this change: b/450351399 is caused by the same flaw in the logic and using the full soaking process that seemed reasonable for the initial bug needs to be sped up.

Followup to ag/35782210, ag/36359572

Bug: 439984935
Test: manual by Android Telecom team
Flag: EXEMPT BUGFIX
Change-Id: If2ff7d2479561a2e181a714c95b050bb066e6228