MirrorView: Use center-fill transform to fully cover the SurfaceView
Now that the SurfaceView sizing has been updated to approximate the
aspect ratio of the session display, ensure that the MirrorView
container is pixel-perfect when scaling down by any fractional amount by
using center-fill transforms to completely cover the SurfaceView with
session content.

Before this CL, it was likely to have a few pixel rows of the
SurfaceView that did not have any content, allowing the background to
seep through. Now, we scale the content so that some pixels of content
may be cropped, while ensuring the full SurfaceView is covered.

Bug: 452461453
Flag: android.companion.virtualdevice.flags.computer_control_access
Test: manual
Change-Id: Ib5b864e6e5f8248abd9d52b76b030c966eedc790