ComputerControlSession: Add package name to blocked state callback
Add a @Nullable blockingPackage param to the onBlocked() callback to
inform the agent of the package name of the app that caused the session
to go into the blocked state, if known.

For secure windows, the blocking package is the package showing the
secure window.

For disallowed activity launches, it is the package name of the blocked
activity launch.

Bug: 442540670
Test: Presubmit
Flag: android.companion.virtualdevice.flags.computer_control_blocked_state
Change-Id: I494b507d1cdebe6345cf474f27dfb99e529434c2