Add resilience to VDM audio
Calling AudioPolicy.createAudioTrackSource() immediately after setting
up the VDM dynamic audio policies can be, in rare cases, racy with
connecting the AudioPorts to the remote submix.
Introduce a resilience mechanism with a few retries for creating the
AudioTrack or AudioRecord for an AudioMix. This avoids system_server
crashes and adds error logs for the failed attempts.
To be followed by a proper syncronization with the AudioFlinger/AudioService
when creating the AudioTrack or AudioRecord, though that is more complex
and requires the right signals from the lower audio layers.

Bug: 452090616
Bug: 451846342
Test: atest VirtualAudioTest --iterations 500
Flag: EXEMPT BUGFIX
Change-Id: Ic1923e5e1ccd8834050ecfee652a7aa4880d9506