Listen to full install signal.
With Archived Apps feature for Restore, apps are first added in pre-archived state, and `onPackageAdded` callback is called. However, apps are not yet installed, and when min aspect ratio is restored via the Settings app, it cannot be restored yet. Another callback has to be implemented, to be able to restore aspect ratio when the app is fully installed, and package info available.

Flag: EXEMPT BUGFIX
Fixes: 454006730
Test: manual, restored user aspect ratio with Archived Apps feature enabled.
Test: atest AppWindowLayoutSettingsServiceTest
Change-Id: I979afe377d0352529910214a55f251f3d172b534