Add a way to instrument pcc processes
This change adds a new parameter that can be used to allow tests to
specify that they wanna run the test in PCC process rather than normal
process.

This aims to make it easier for any restrictions that we enforce on pcc
processes to be tested.

Bug: 438428601
Test: atest CtsPccProcessTests
Flag: android.app.privatecompute.flags.enable_pcc_framework_support
Change-Id: I0d4097d50c242dd173870852da4cc29e817faf83