[pm] Refactor: Move package update logging to InstallPackageHelper
Move the logging for a package code path update from Settings.java to InstallPackageHelper.java.

This change delays the printing of the logging to the actual time when
the package setting is committed, thus making it easier for debugging
timing related issues. The new log also prints the actual final code
path instead of the temporary code path.

BUG: 390566126
Test: manual
FLAG: EXEMPT PURE_REFACTOR

Change-Id: I181c8d65177d894e4144f6a12f8e4f3bc44cd9a6