Prevent connecting to mSocket multiple times
This commit fixes a bug ag/35806022 introduced - connect() could be
called on mSocket multiple times. The fix was in ag/36017455 and the
original plan was to submit the CLs at the same time, but as only
ag/35806022 has been merged, we create this separate commit to fix it.

Test: atest NativeZygoteIntegrationTest
Bug: 442732151
Flag: android.os.native_app_zygote
Change-Id: I889d1e991fac9b4f5ab2163bb3bbe9240308e356