Create UserResolver interface
Initial check in of an interface for resolving user info needed for
displaying bubble icons for different profile types. This initial
implementation uses UserManager to resolve the user type which makes
a binder call. In the future we should introduce caching and update
the interface to a suspend function.

Bug: 443920865
Flag: EXEMPT BUGFIX
Test: atest UserManagerUserResolverTest
Change-Id: I324d280e0e431daaa13522a90d155d6bf2001460