Send profiling trigger on excessive resource usage kills.
This change sends a profiling trigger when a process is killed for
excessive CPU usage. This allows the profiling service to collect data
on misbehaving apps.

The trigger is sent when a process is killed for:
- Excessive CPU usage.
- Excessive binder traffic that prevents the process from being frozen.

Test: Sample app runs in high CPU usage, and confirm the trigger
received from Profiling Service
Bug: 448360513
Flag: android.os.profiling.profiling_trigger_kill_excessive_cpu_usage

Change-Id: Ie544644fc253921501da517b5d0202f6d7dc2ed8