Make BinderProxy frozen callbacks more debuggable
Originally frozen callbacks are wrapped with an inline lambda that
invokes a callback on the associated executor. However this wrapper
makes it hard to see the original callback class names in crash logs.

This change removes the wrapper, and instead stores the executor in a
callback -> executor map. The original callback is directly passed into
the native binder code and therefore its class name would now be visible
in crash logs.

Bug: 454431919
Test: TH
Flag: android.os.binder_proxy_use_executor_map
Change-Id: I7b3797fdf0ddb541c37bc8f064df625e64b784f3