Add ConnectivityCallListenerService to the manifest
This change adds the ConnectivityCallListenerService to the Android manifest. This service is an InCallService that allows the Connectivity module to monitor call connectivity state.

The service is configured to include self-managed calls, which is necessary for it to be aware of calls from third-party VoIP apps so that it can route its traffic over premium network.

Bug: 430509489
Test: atest InCallControllerTests, EnhancedConfirmationInCallTest
Flag: EXEMPT BUGFIX

Change-Id: Idf69b174bf701bff9897e453d783deb24df38124