Hardcode behavior of polish_close_wallpaper_includes_open_change
This change removes the runtime check for the
`polish_close_wallpaper_includes_open_change` flag in ActivityRecord.java, effectively hardcoding the behavior to its enabled state.

The flag is fully enabled on the main branch but remains disabled on the 25Q4 branch. This difference causes
`ActivitySecurityModelTest#testTopFinishesThenLaunchesActivity_launchAllowed` to fail on the 25Q4 branch.

This CL is intended to be cherry-picked to the 25Q4 branch to fix the test failure. Hardcoding the behavior on main provides a minimal, targeted change for the cherry-pick, posing less risk than enabling the entire feature flag on the release branch.

The flag definition will be completely removed from the main branch in a separate, follow-up change.

Bug: 429317375
Test: atest CtsWindowManagerBackgroundActivityTestCases:android.server.wm.ActivitySecurityModelTest#testTopFinishesThenLaunchesActivity_launchAllowed
Flag: EXEMPT BUGFIX
Change-Id: I53edbbe533831cf1798cd353df8d286e046ca2fc