[Flexi] Always compose the shade scene.
On the first frame of shade expansion, the entire shade scene needs to
be re-composed, which takes an incredibly long time and causes jank.
This CL changes the shade scene to be always composed, so that the
entire scene isn't re-composed on the first frame of shade expansion.

This CL includes alwaysCompose=true plus a few fixes for issues caused
by always composing the shade:
 - Tile squishiness incorrectly propagating for dual shade: Changed to
   LaunchedEffectWithLifecycle (b/458108271)
 - With dual shade enabled, notifications shade is incorrectly showing
   on restart: Updated ShadeSceneContentViewModel (b/458102847)
 - StickyHeadsUpIsCancelledAfterLongWait failing on desktop
   (b/458108266): Same fix as above
 - QQS tiles not marqueeing: Passed correct `listening` value

Fixes: 410757909
Flag: com.android.systemui.scene_container
Test: atest SystemUITests
Test: Smoke test of shade expand & collapse
Test: Smoke test of notifications
Test: Smoke test of quick settings tiles
Test: See comment42 for more information on testing done.

Change-Id: I959573f3d28eefd63d84b323277c10eaf6ee7a1c