Provide desktop state visibility through HomeTransitionObserver
Updates HomeTransitionObserver to include desktop visibility information
when home is shown behind the desktop. Taskbar uses home visibility
state to determine when in-app taskbar is shown - this doesn't work
great in desktop mode when home is visible behind desktop windows, and
taskbar ends up thinking that it's not in-app state when in desktop
mode (more evident in tablet mode - in desktop first mode, taskbar
is visible on home, so knowing when in-app state is active is not as
important). This updates HomeTransitionObserver, which is used as source
of truth for home visibility by taskbar, to provide information about
whether home is visible behind desktop, so taskbar can use that to
determine "in-app" state on devices that show home behind desktop
wndows.

Note: I was considering using "active desk" signal to determine whether
home is behind desktop windows, but that does not handle the case where
an empty desk is active (and home is visible with no desktop windows
on top), plus the approach here may be more flexible when it comes to
timing with different transitions.

Bug: 474391544
Bug: 446091472
Test: manual
Flag: EXEMPT BUGFIX

Change-Id: Ib8f00e3f3733f9921eca5389b15e15d85cd1d5c4