Mark uses_cleartext_traffic deprecated, add (disabled) deprecation logic
The new behavior is not enabled, it's a @Disabled CompatChange to let developers test locally. In a future release, we'll enable this change.

This change:
- marks the XML attribute usesCleartextTraffic as deprecated (this part cannot be flag-gated, as per offline discussions)
- marks isCleartextTrafficAllowed() in AndroidPackage.java as deprecated (flag gated)
- marks isCleartextTrafficAllowed() and setCleartextTrafficAllowed() in PackageImpl.java as deprecated (flag gated)
- updates the logic in ManifestConfigSource.java to force usesCleartextTraffic=false if the flag is enabled
- adds a unit test to validate the change in ManifestConfigSource

Context: usesCleartextTraffic=true will be deprecated in favor of Network Security Config Files, which offer the same capability but with more fine-grained control, in particular, most apps should only allow cleartext traffic to some small set of domains if they need it. More context in go/deprecate-usescleartexttraffic-bluedoc .

Test: atest NetworkSecurityConfigTests:android.security.net.config.UsesCleartextTrafficDeprecationTest

Bug: 415007211
Flag: android.security.deprecate_uses_cleartext_traffic
Change-Id: I08746468b4f0fe8e94a29dc417120a151cae0d9a