Consolidate Duplicate Code
This change eliminates the duplicate histogram bounds array that was
defined in both RelativeFrameTimeHistogram and PendingJankStat classes.
It also streamlines the accessing of histogram buckets by adding a
static method to RelativeFrameTimeHistogram for accessing bucket
indexes.

This change also moves the merging of the histogram logic into the
PendingJankStat class.

Bug: 450104820
Test: atest CoreAppJankTestCases
Flag: EXEMPT PURE_REFACTOR

Change-Id: I33cdbb89872361e84755df1adffd2e1b6be6b569