PowerCommand: Schedule wakeup from "android"
cmd power wakeup {time} currently schedules a wakeup on behalf on shell
but it's observed that when shell process is frozen by ActivityManager,
this alarm will be removed and devices will never get to wake up. The
process freezing could happen when devices are on battery or when screen
turned off right after boot after since there is a gap between
PowerManagerShellCommand is ready and ActivityManager receives idle
policy after booting.

Schedule on behalf of "android" system server process instead as this
process won't be frozen.

Bug: 455634740, 451857479
Test: update-device & reboot & immediately cmd power wake up 20000 & cmd power sleep, wake up will happen even though log suggests shell is frozen.
Flag: NONE test only
Change-Id: I2fcf0d049759e88d2dd3ab52f7972841ec585323