AppHandle: Always force show system bars.
There should not exist a case where we want the handle menu open but the bar disappearing, so will always force show them. This is mostly for the case of immersive apps that auto hide bars after awhile - we want to ensure they don't auto-hide until the user dismisses the menu.

Test: Manual - open menu while in an immersive app, see system bar stays visible
Bug: 375253742
Flag: EXEMPT BUGFIX
Change-Id: Ib6df405069eb557aa12b4257c9928df7fc8e4fc6