[PiP on Desktop] Fix expand for multi-activity PiP in Desktop overview.
Previously, we added a workaround for a separate bug to always expand
PiP to fullscreen when we are in Overview initiated while in Desktop
session. This workaround does not work for multi-activity PiP tasks,
because even if we are setting the windowing mode of the PiP task to
FULLSCREEN, its parent might still have FREEFORM windowing mode.

DesktopPipTransitionController#maybeReparentTaskToDesk is supposed to
update the parent's windowing mode appropriately, but it was mistakenly
setting the windowing mode to FREEFORM as long as Desktop session is
active, rather than using PipDesktopState#getOutPipWindowingMode which
has the fullscreen in Overview workaround.

This CL updates maybeReparentTaskToDesk to use getOutPipWindowingMode.
Also cleans up the method to make it more readable and simple.

Bug: 455666441
Test: Manual - expand PiP while in Desktop overview and verify it
expands to Fullscreen
Test: atest PipDesktopStateTest DesktopPipTransitionControllerTest
Flag: EXEMPT BUGFIX

Change-Id: Iae0823f94b2c83c53f3fa8a84c00f1a2d4695319