Wait for app process to die when updating the package
Currently, during update we simply kill the app process in async manner
without waiting for the process to die first.

We now need to wait since apps may take up to 11 seconds to dies during
package udpate in some cases. This is because WindowManager may ask the
app process to save their state during update.

We will be callling a new stopAndKillApplication API and wait for it.
However, we can't block the main handler thread. The waiting is
transferred to a background thread and installation is resumed on main
thread once waiting is complete. This requires we turn
commitInstallPackages async.

The feature is guarded by aconfig flag. When the flag is enabled, we
will wait for package to be killed during package update. When disabled,
we will not wait and assume package is killed instantly (existing
behavior).

Bug: 454290193
Flag: com.android.window.flags.enable_app_restart_after_update
Test: atest CtsStopAndKillHostTest
Change-Id: Idf3ec5e1ef2ea36f840ab74147590f33911ff70b