Add Uprobestats bridge service to classpath
This change integrates the Uprobestats bridge service into the
system by adding its bootclasspath fragment, which is delivered
via the `com.android.uprobestats` APEX.

The service's inclusion is conditioned on the
`RELEASE_UPROBESTATS_BRIDGE_SERVICE` build flag.

In `SystemServiceRegistry`, the service's initializer is now
called, gated by the `uprobestatsBridgeService` runtime flag.

To support building `framework-base` when this flag is disabled
(and the APEX is not present), a mock
`UprobestatsFrameworkInitializer` is added via a new genrule in
`core/java/Android.bp`.

Test: atest CtsDynamicInstrumentationManagerTest
Bug: 412739902
Bug: 381414268
Flag: android.security.uprobestats_bridge_service
Change-Id: I181e7270af29726045cf93b77673424d59f2dde0