Make SystemEventChipAnimationController per-display
Before, the MultiDisplaySystemEventChipAnimationController was being
used, which is a @SysUISingleton, and delegates calls to the
SystemEventChipAnimationControllerImpl of each display.

Now, we simply make SystemEventChipAnimationController be called from
the classes of the same display only.

Flag protection is on the dagger level. Dependencies are Lazy, to avoid
unnecessary instantiation when the flag is on/off.

StatusBarWindowController is now temporarily a nullable dependency in
SystemEventChipAnimationControllerImpl, to avoid NPE during Dagger
initialization, for the cases when the associated display is destroyed
during Dagger initialization.

Flag: com.android.systemui.system_status_animation_per_display
Test: Connect a display and start/stop audio recordings
Bug: 444658781
Change-Id: I29e12abbd0e30d58e76ec8677efdfd44a5e40f1d