[Flexiglass] Improve Notification transitions between AOD and LS
Notification AOD icons, and Notifications are never meant to be displayed together on AOD. Ideally we would only display here one, or the other, but keeping the Notification around at the same position, where it needs to animate in from helps with the NSSL Flexi integration.
This CL makes the NotifPlaceholder and NotifIcons overlap on AOD.

BurnInModifiers aggregates movements of elements to/from AOD. These modifiers are applied one-by-one to elements, because there are some that don't actually need them.
This CL applies them to the Notifications in the upper region of the LS, AOD.

Fixes: 454605883
Test: Navigate between AOD and Lockscreen -> check the appear animaton of Notifications
Flag: com.android.systemui.scene_container

Change-Id: Ie1616256130a2204e62186e2356826ccb52b341a