Lazily list APEX certificate directory
Prior to this change, the complete content of the APEX directory was
enumarated to confirm that it contained at least one element. This had
an impact on the memory footprint of apps, as well as their startup
time.

Use lazy evaluation to ensure that the directory contains at least one
entry.

Bug: 456298201
Test: presubmit
Flag: EXEMPT BUGFIX
Change-Id: Ie8c60232839828d26938124e97d32a66e1920e75