[AOA FFS Implementation] Update Usbdevice Manager to enable CTRL
This change handles the CTRL function in the usbdevicemanager similar
to ADB. In case of ADB it is added to functions based on if ADB is
enabled or not. In CTRL we will always add it to the functions.

This change also handles the calling of updateAccessoryState from the
native code. We call it outside the block so that lock can be released
when the upcall to Java is made.

Bug: 457946796
Test: Tested locally with the WebAoa application
Flag: android.hardware.usb.flags.enable_aoa_userspace_implementation

Change-Id: I41f94d19162703c441db5f414b84d4d2cbc01077