Fix a subtle issue with bubble removals & overflows
When the stack is dismissed, the group of bubbles in it
get overflowed -- so dismissal (removing the views from the stack)
and overflowing (adding overflow bubbles) happen together.

This is usually handled fine, but not in a unique case where the
one of the bubbles being dismissed/overflowed might cause the removal
of one of the other bubbles being dismissed/overflowed.

When something is removed from the stack, we clean up the views.
We also do this when removed from the overflow. If the overflow
does it first, the removal on the stack doesn't have a view to look
up to remove from the stack.

We already clean up the views when removing from BubbleStackView and
BubbleBarLayerView, so this call is unnecessary, so we can just remove
it. Existing tests for cleaning the views up exist in
BubbleStackViewTests, added a new one in BubbleBarLayerViewTest

Flag: com.android.wm.shell.remove_app_bubble_overflow_dupes
Bug: 415103355
Test: atest WMShellUnitTests BubbleBarLayerViewTest
Change-Id: I2a9a2a150b7e3e2fb6275fdce4aab9b510a67cb8