Handle app content service binding death
The `AppContentProjectionService` in target applications can
crash or be killed, leading to a `ServiceConnection` that never
receives `onServiceConnected` or `onServiceDisconnected`.
This change overrides `onBindingDied` in
ScreenCaptureAppContentRepositoryImpl's `ServiceConnection` to
gracefully handle such scenarios. When a remote app content
service dies, `onBindingDied` will log the event, unbind the
service, and emit a `Result.failure` to the flow. This prevents
the SystemUI from hanging indefinitely while waiting for a
response from a crashed service, improving robustness and user
experience.

Flag: com.android.systemui.large_screen_sharing
Bug: 423708479
Test: m
Change-Id: I8bcfe4b9b49cafe651d47165e1f2e4a87f2d6648