Send pending events to non rr subscribers when not in foreground
We want to not send refresh rate changed callbacks when the client is in the background. However the current implementation drops out all the pending events. This bug was recently introduced when we bacthed the events to an eventmask approach

Bug: 449949226
Test: Manual
Flag: com.android.server.display.feature.flags.send_non_rr_callbacks_when_in_background
Change-Id: I7a6232d04199b039a6aed297157f998ca0db3ea3