Remove bindIme from setImeVisibleOrReconnect
In [1] a call to bind the IME was introduced in setImeVisible, if we
receive an IME show request while no IME is bound. However, trying to
bind at this time could likely fail if the show request comes after
resetSelectedInputMethodAndSubtypeLocked, but before a startInput, as
the mSelectedDisplayId would be INVALID_DISPLAY, which would fail to add
the WindowToken for the IME, as well as fail to add the Window for the
IME.

Instead, we should only bind when we receive a successful startInput, as
that computes the correct displayId to show the IME.

One such failure case is full user switching on slow devices. The next
user's Launcher could be started before the user switch completes. At
this time the IME Control Target is still the previous user's Activity.
If that had the IME requested visible, we would try to show it again,
while no IME is bound, and no startInput call was received (before
Launcher handles the input focus change).

  [1]: Iacb29cfdb3825248ee8c40e610652f697aa2d6da

Flag: EXEMPT BUGFIX
Bug: 281029564
Test: CtsInputMethodTestCases
Change-Id: Icee10231402603a278b879ab035e4fa4824458f9