Add API to handle missing route permissions
Add a setPermissionsComponentName API to RouteListingPreference.Builder.
This allows apps to specify an activity that can handle requesting
permissions when they are necessary for the app to see some routes.
The primary use-case is for apps missing the
ACCESS_LOCAL_NETWORK permission.

When a RouteListingPreference is specified with this option, the
MediaOutputDialog will show a warning message and a link to the provided
component, so users can grant permissions to the app.

The warning message is only shown if:
 - There is a RouteListingPreference with the component set
 - There are routes that the app does not have access to
 - The app has requested permissions that would be sufficient to see the
   routes if they were granted.

Test: atest SystemUITests CtsMediaRouterTestCases (see changes in topic)
Flag: android.permission.flags.access_local_network_permission_enabled
Bug: 459390619
Change-Id: Ic9138b5987ce375fcfe35efc557d9eff0e1c614b