Return app_start_info_process_name_fix flag back to its original state
A subsequent change had accidentally removed the BUGFIX metadata from this flag declaration, which caused problems with the flagging system. This puts this flag back to the way it originally was declared (which is what the flagging system wants).

Bug: 447677155
Test: n/a
Flag: com.android.server.am.app_start_info_process_name_fix
Change-Id: I4f5e945ed82e25cd6f669473e1f618bb890c5708