Initialize ProtoLog asynchronously in ActivityThread
To improve the performance on the caller thread. With the caveat that we
might end up tracing more data the necessary once ProtoLog initialized
fully. The async implementation queues up messages with the correct
timestamp (at call time) to be traced later when the Perfetto datasource
and producer get registered properly, until when we don't know the exact
tracing configurations that may be running so we keep track of all
message to post them to the datasource once initialized fully. Worst
case we trace messages that were posted after a datasource instance was
started in the racy case (should be pretty race), so shouldn't have
technically been traced cause they were traced too early, but in
practice this should be a non-issue, we will never miss data we would
have had otherwise if using the synced initialization.

Flag: android.tracing.protolog_async_init
Test: atest TracingTests
Bug: 434072453
Change-Id: Ia88507dad35492d99985fdf16a0fa224c6aa6639