Enable bubble task reparenting on home relaunch.
This change applies the setReparentLeafTaskIfRelaunchFromHome
WindowContainerTransaction API to the bubble root task.

This allows the window organizer to control whether a leaf task should
be reparented to the task display area when it is relaunched from the
home screen. This provides finer control for bubble relaunch scenarios,
such as moving a bubbled app to fullscreen when its launcher icon is
clicked.

Bug: 456621541
Flag: com.android.window.flags.enable_bubble_root_task
Test: atest WMShellUnitTests:BubbleRootTaskTest
Change-Id: I40bc6a95fe93dde27edccbdac5fabc696552e4b0