Prepare for Metalava becoming stricter with syntax errors
Metalava currently only reports syntax errors at the class level which
means it can ignore syntax errors at the class level. That includes
ignoring class modifiers that are not supported by the java language
version. e.g. using `sealed` on a class that is for a java language
version less than 17. A future Metalava change will fix that.

This change changes the java language version to "17" as some of the
sources use the `sealed` modifier.

Bug: 355648520
Test: m checkapi
Flag: EXEMPT BUGFIX
Change-Id: I2031b06517c304cfbae33548b0a9886d2db891f8