Rename NSSL stackTop to stackScrollTop
This field is a critical link between the composable placeholder that is
being scrolled, and the legacy NSSL that places its children.

This CL adds `scroll` to its name to help to distinguish it from other
fields like `stackSpace.top`, which are not offset by scrolling.

Bug: 457958356
Test: atest StackScrollAlgorithmTest
Flag: com.android.systemui.scene_container

Change-Id: I21c984f22b99c0cf8bc3437cf19463a64255118f