Rename showGlobalActions to showOrHide
The method called at the endshows or hides the dialog. This is used by
PhoneWindowManager to show/hide the power menu on long press of power
button. This change only renames the intermediate methods to reflect
this behavior.

This will facilitate adding a showGlobalActions method in a follow up
that does is a noop if the dialog is already showing.

Test: build
Test: atest StatusBarManagerServiceTest
Test: atest SystemUITests
Flag: EXEMPT PURE_REFACTOR
Fixes: 458719236
Change-Id: Iaa7c70ae9d9e241d3da92e8996929b5a123fd037