Don't transition to lockscreen/gone when swiping up on the dream
When swiping up on the dream, that will cause the dream to terminate. After the dream terminates - it will naturally transition to the right scene. If we start transitioning as you swipe, you can get into a state where the lockscreen content is visible on top of the dream before the user lifts their finger to end the swipe

Bug: 458435039
Test: manually on device by swiping up on the dream to exit
Flag: com.android.systemui.scene_container
Change-Id: If7e6187aaeddf06adf1618db6cac7deddb131195