Fix bouncer delay for biometric auth during drag/fling transitions
The current delay to show the bouncer (which is applied when a delayed
authentication may succeed) is only applied during non-user initiated
transitions to bouncer, e.g. when pressing on lock or notifications.
This is fixing this by combining both combinations, the automated and
the user initiated one during drag/fling transitions.

Bug: 458616685
Flag: com.android.systemui.scene_container
Test: manual opening bouncer with face auth enabled
Change-Id: Ia3d7f4d03b4e42002429dc415a6024a5d8e98a59