usb: userspace aoa: do not error check in protocol handlers
The Android Open Accessory protocol has historically not had proper
testing and validatation, and therefore there are headunits which do not
follow the protocol exactly. To maintain compatibility, do not fail to
negotiate due to behavior which is not per the spec.

Further validation of the logic is required when compared to the kernel
implementation, but this resolves a known issue.

Bug: 407985367
Bug: 416222959
Test: Verify the connection is not reset with android auto headunit that
sends start accessory with value set to 1: "Malformed start accessory:
value: 1, index: 0, length: 0"
Flag: android.hardware.usb.flags.enable_aoa_userspace_implementation
Change-Id: Iea2f4bd7a151fbeb0855b25f7387978bc311b6ed
Signed-off-by: Neill Kapron <nkapron@google.com>