Reapply permissions, component check refactor
Factor out code to check for activity components and route permission
checks to their own methods. This will be used in a followup to verify
an additional component in RouteListingPreference, and to check
permissions for ManagerRecord to check permissions of its target
package.

Note a previous version of this change was reverted to inverted logic
(!= PERMISSION_GRANTED instead of == PERMISSION_GRANTED).

Test: atest (existing tests)
Flag: EXEMPT PURE_REFACTOR
Bug: 459390619
Change-Id: I24ac944b595d5622c81bfd2159e509b6ac16dc6f