[Flexiglass] Fix crash during Shade to QS transitions
When navigating from SingleShade to QuickSettings with a strong (high velocity) fling, the NotificationScrim is moved down so far, that it crosses the Ruler at the bottom of the screen, which we thought will never happen. This tries to create a constraint with a negative height value, which is not allowed, and it is causing a crash.

This CL constraints the calcualted content height to be at least zero, so we would place a 0 height content when this happens, but not crash.

Fixes: 458624498
Test: Fling down in SingleShade -> QuickSettings opens
Flag: com.android.systemui.scene_container

Change-Id: I418c418d6ad770af1b906d588fc7cad4ff89c616