Don't send display events if no access
A process should only receive display events for displays that it has access to.

After a display is removed, we don't do the access check because the process is no longer present on the display and the check would return false. We send the event to DisplayManagerGlobal and then use the cache to see if the corresponding added event was sent.

Bug: 458435043
Flag: com.android.server.display.feature.flags.display_ids_cache
Test: DisplayManagerServiceTest, DisplayManagerTest, DisplayEventDeliveryTest
Change-Id: Ic380e949d2095dd095a33431c5dc0943c69c86bd