[flexiglass] Defer PIN bouncer input shape loading
This defers the loading of the shapes for the input area / text field of
the PIN bouncer until after the bouncer overlay transitions in
completely; which is a performance optimization done to prevent the
long-running operation of loading the ~10 or so animated vector
drawables right as the user is entering the bouncer, which causes jank.

Fix: 453647501
Test: added logging statements to SceneJankMonitor to print something
when beginning or ending a CUJ and to right before the expensive
animatedVectorResource call. Saw that it's loading the shapes inside the
CUJ. Made the code changes and did it again, proving that the loading
now happens only after the CUJ ends.
Flag: com.android.systemui.scene_container

Change-Id: I730c2f3fb5ea8eba6ecd6fe770c67e01de06049d