am: Encapsulate service start/stop in ProcessServiceRecord
This CL refactors the service start and stop logic to improve
encapsulation.

Previously, ActiveServices would call into ProcessStateController to
add or remove a ServiceRecord from a ProcessServiceRecord. This makes
ProcessStateController depends on the implementation details of
ProcessServiceRecord.

This CL changes the call sites in ActiveServices to instead call
startService() and stopService() directly on the ProcessServiceRecord.
These methods then delegate the actual bookkeeping to new methods in
ProcessStateController (`addRunningService` and `removeRunningService`).
This makes ProcessServiceRecord the authoritative entry point for
managing its own state.

Bug: 457589118
Test: m services.core
Test: atest MockingOomAdjusterTests ActivityManagerServiceTest
Test: atest FrameworksServicesTestsRavenwood_ProcessStateController
Flag: EXEMPT PURE_REFACTOR

Change-Id: I03a957aed28ec6b1018fa813511b40a5ff92bf7f