Capture only required properties of parent for nested lockscreen scopes
Some aspect of the way NestedSceneTransitionLayout seems to require that
we don't capture the parent ContentScope in the builder callback. When
this happens, it appears to hold the reference to these lambdas much
longer than expected.

I'm unsure of the exact mechanism here, but we don't need to capture the
parent scope regardless. Addtionally, if this doesn't fix the leak, it
will still minimize the impact of the leak by only capturing the scope
objects that are strictly nessecary for building the child scope we
require.

Bug: 459473782
Test: Checked nested scope behaviors
Flag: com.android.systemui.scene_container
Change-Id: If8dcb1d32635e40f540c031a90b866964f3b9da0