Fix ToggleShadeWithKeyboardShorcut on older devices
ToggleShadeWithKeyboardShortcut was  failing on older devices due to the
shade not opening on the overlay display. The logic for the keyboard
shortcut opens the shade on the focused display, and the focused display
policy only updates focused display when a device is capable for desktop
mode (which all P22 or earlier devices are not). To remedy this, now if
desktop mode is not enabled, the shade will simply be opened where it is.

Fix: 453909296
Test: ToggleShadeWithKeyboardShortcut (on felix)
Test: SysUIKeyGestureEventInitializerTest
Flag: com.android.systemui.shade_window_goes_around
Change-Id: I69c8ff20f1cf556a8cd8063f43117a4b4e844472