Auto-close recovery sessions after a minute.
Recovery sessions shoudn't stay alive indefinitely, and so we should
automatically close them after a reasonable timeout period. This isn't
strictly a security feature as the timing of the recovery operation is
enforced by the HSMs and so we're not relying on this to prevent
long-running sessions.

This implements a fake scheduled executor service for use in testing.
It only supports the features we currently make use of (so Runnable but
not Callable, and only "schedule") but it would be easy enough to fill
in the other features if we wanted to use them.

Bug: 444216115
Test: mmma, atest LskfResetTests
Flag: android.app.lskfreset.flags.enable_lskf_reset_manager
Change-Id: I0f9237bcf054ec8009b7edfff4a16c3504ee91c8