[flexiglass] Skip recomposing PasswordBouncer too much
Each time the bouncer overlay is shown, in the password configuration, a
second composition happens quickly after the first. This is introducing
jank into the relevant CUJ.

The second recomposition happens because the first composition reads
from the isFocusRequested state from the view-model and that state
rapidly changes from its initial false to a true, soon after the
view-model is activated.

While it seems safe to assume that it's okay to start with the text
field focused, that assumption isn't actually safe as we need to support
the disabling of the UI when the user has entered the wrong password too
many times and is being "throttled"; something that can happen while the
user in looking at the bouncer or when the user returns to the bouncer
after having left it during the throttle countdown period.

The solution presented here relies on strongb skipping in Compose. It
moves the part that reads that state off to a child composable function
which means that only that child composable is recomposed when the state
changes.

Fix: 456075665
Test: manually verified that a Log.d statement in the first line of the
top-level composable is only printed once after the change (twice before
the change)
Test: manually verified that focus is re-requested when the throttle
countdown finishes and the UI is re-enabled
Flag: com.android.systemui.scene_container

Change-Id: Icda2e04b076bbb577d154742f18cf76294049a88