Fix TaskDisplayArea.mRootPinnedTask not being set in external displays
when PiP is dragged to a new display. This caused a regression in
drag-PiP-across-displays, since RootWindowContainer#getRootTask can't
find the root task for WINDOWING_MODE_PINNED, which led to the Settings
button on the PiP window being non-functional.

The regression was caused by the removal of the line
addRootTaskReferenceIfNeeded in TaskDisplayArea#addChildTask in the CL
ag/36104684.

Bug: 458152889
Test: atest TaskDisplayAreaTests TaskTests
Flag: EXEMPT BUGFIX
Change-Id: I7c5df742ab252133c3db389cbaf57a18b113ebb9