Bring up preshareUI when launch ShareScreenActivity with user 0
go/multi-user-and-hsum-in-systemui suggests to always launch UI activity
in user 0 instead of current user. Based on this best practice, we are
migrating to launch ShareScreenActivity with user 0.

One issue with launching ShareScreenActivity with user 0 is: the
preshare UI cannot be bring up as it runs on user 0, not visible to
current user. This change fixes this issue by explictly telling
WindownManager, even the window belongs to system user, display it on
the screen of the current foreground user.

Bug: 459901401
Bug: 456823719
Flag: com.android.systemui.large_screen_sharing
Test: manual
Change-Id: I32326f57ba5bd24dc7a69566c0caaf2bacccd1e8