Fix PipelineCacheStore race condition on exit
There are two issues with the current implementation:
 1. As `mExit` is modified without any synchronisation (via compare and
    exchange), it is possible that the thread observes `mExit` as false
    and then goes to sleep while being notified to wake up, missing the
    wake-up notification.
 2. As there is no release / acquire ordering between `notify_one` and
    the wake-up, it is possible for the thread to wake up and still
    observe `mExit` as false.

Both issues cause the two threads to deadlock until the next spurious
wake-up.

To resolve both of these issues, demote `mExit` from an atomic to a
normal variable, and protect the write to it with the mutex instead. Due
to the mutual exclusion, the notification can't be missed, and due to
the release / acquire ordering from the atomic variable in the mutex,
the thread is guaranteed to observe `mExit` as true when it is woken up
by `notify_one`.

Bug: 269117286
Test: boot
Flag: com.android.graphics.hwui.flags.separate_pipeline_cache
Change-Id: I4378d1bf81448ca23d6e0bc27676d5b3d11a3b85