Add PiP bounds change transition jank cuj tag
Add a jank CUJ tag for CUJ_PIP_TRANSITION to keep
track of bounds change transitions when interacting
with PiP.

More specifically, every time a resize transition is being
played for PiP BOUNDS_CHANGE_TRANSITION tag is started, ending
when the transition finishes.

Note that this doesn't cover the undeterministic PiP interactions
which are animated differently and should be tracked separately:
these are flings and the pinching of PiP pre-release, where the user
hasn't yet committed to a bounds change transition.

Bug: 352738878
Flag: EXEMPT BUGFIX
Test: double tap to resize with
	InteractionJankMonitor instrumentation enabled
Change-Id: I64766f4c97e25e8239a95ea2fa8848d61b9937c7