[Ravenwood] Make ContentResolver mock-able
Keeping several methods in ContentResolver class is necessary for any
mocks of ContentResolver to work properly due to how ContentResolver is
implemented. Add a detailed comment in the annotation to let users of
this class understand that even though many methods appears to be
supported, it does not actually work the same as it does on device.

Bug: 457841012
Flag: EXEMPT HOST_MODE
Test: atest FrameworksCoreTestsRavenwood
Test: atest RavenwoodCoreTest
Test: f/b/r/scripts/run-ravenwood-tests.sh
Change-Id: I148cfc306166b80a92d4b72287ac8b293abc153d