Correct kotlin styling in selection toolbar
This change applies more standard styling practices to the selection
toolbar code that was converted from frameworks Java classes. This
change includes changing naming conventions, defining fields in
constructors, fixing nullability notations to be appropriate in the
context, and using idomatic expressions to improve readability. The
biggest contributor to lines of code is that the init {} block was
removed entirely from the RemoteSelectionToolbar class, field
declarations had to be reordered to be consistent with the order the
Java constructor initialized the class.

Bug: 450979790
Test: Manual for now
Flag: com.android.systemui.use_systemui_views_for_system_toolbar
Change-Id: I94607a2e3ec568b9e1a73649800b2d44c8ae39d3