Fix A11yIME Session lifecycle
Because session is created per client, every time the client changes,
sessions with old client must be disabled and sessions with a new
client must be enabled. However, the current implementation for a11y
IME session is broken, and until the a11y service is restarted, we
never re-enable the session.

Bug: 454700008
Test: AccessibilityImeTest
Test: AccessibilityInputMethodTest
Flag: EXEMPT BUGFIX
Change-Id: Iabd40b91c0dde81ab6fdff5fd348801962997268