Add IME subtype short label API.
The short label is a non-localized string used to represent the subtype
in compact UI elements where the full display name is too long. As an
example, it could be an abbreviated locale code.

The short label should contain at most two characters since it may be
used in place of an icon in some UIs if a subtype icon is not specified
(e.g. on an IME chip to indicate the selected subtype).

Bug: 443854916
Flag: android.view.inputmethod.ime_subtype_short_label
Test: atest FrameworksInputMethodSystemServerTests:AdditionalSubtypeUtilsTest
Test: atest CtsInputMethodTestCases:InputMethodInfoTest &&
  atest CtsInputMethodTestCases:InputMethodSubtypeTest
Change-Id: Ie4c49b96aa1def0491a6060ec53b03211cdbcae4