[Flexiglass] Re-enable ShadeSceneTest
ShadeSceneTest was disabled because it failed to activate SingleShade when DualShade was used by default.
This is a problem that was mitigated for other SingleShade unit tests since, and the treatment is to disable DualShade's aconfig flag + call the kosmos helper to enableSingleShade().

The CL applies the patch, and removes the @Ignore annotation from ShadeSceneTest.

Fixes: 425670368
Test: ShadeSceneTest
Flag: com.android.systemui.scene_container
Change-Id: I000d43763837311eb89f1b02c75d29c4187ef017