Make shade window flag behavior match legacy behavior
Prior to the scene container, the window was focusable as soon as
the shade began to expand, and the current behavior is to make the
shade window focusable when expansion completes. There is concern
that this will cause hard to diagnose bugs at some point, so this
change will make the behavior match the old behavior.

Fixes: 458634569
Flag: com.android.systemui.scene_container
Test: updated unit tests
Change-Id: Ic82972efb1f21a9eb83f8113b7a0ecd50551ae89