Clarify ACTION_SHUTDOWN is per user
ACTION_SHUTDOWN's documentation is ancient and speaks of the "device".
But it is actually sent at the shutdown of each user when that user
stops. It is somewhat the counterpart of BOOT_COMPLETED, which already
speaks of "user" rather than "device". So we update the documentation to
align with other "user" documentation.

Fixes: 456254307
Flag: EXEMPT DOCS_ONLY
Test: N/A

Change-Id: I87b5d610844ee1737ff10bd7af489d67684f2d98