[mobile] re-add `requestLayout` to stacked mobile icon
The change with id: Ibdc0a61f402bf7b515a333ff1b738533f1b50268 fixed the
issue with StatusBarRoot taking up the max width, but it seems that
removing `requestLayout` from StackedMobileIconBinder was not necesary
for that fix and allowed for layout issues to persist.

This CL re-adds the requestLayout to the stacked mobile icon so that
there is no clipping happening.

Test: with 2 compatible SIMs (i.e., stacked mobile icons), expand quick
settings once. So that the stacked icon shows above QS. Turn on airplane
mode, let the UI settle, and then turn airplane mode off. The icons
should show with the RAT indicator next to them.
Bug: 449043701
Flag: EXEMPT BUGFIX

Change-Id: Ia4613240dda2592eee313161c31dc848f0145366