Use pill buttons for RONs and FSI
* RONs, FSI, and CallStyle use pill buttons with common general logic but with different defaults.
* RONs, FSI, and CallStyle use evenly-divded notifications (all the way to the left gutter).
* CallStyle uses filled action buttons (always). FSI uses outlined action buttons (always). RONs respect emphasis set in Actions (except that if there's only one action, it's forced to outlined).
* CallStyle uses text and icon (always). FSI uses text only (always). RONs respect style set in Actions (icon only, text only, icon and text).
* In RONs, FSI, and CallStyle, the semantic color is applied to the button (either background or outline) if it's a full-length span (same as for ForegroundColorSpan).
* Changed some contrast calculations to account for the different color usage.

This CL forks applyStandardTemplateWithActions and related methods to ensure minimal disruption to the unflagged path (both the structure and behavior has changed significantly). The only change to the legacy path is the extraction of displayRemoteInputHistory(), which is unaffected and makes sense as a separate method anyway.

Added a mask to btn_notification_emphasized so that the ripple always extends to the full button (even if it's transparent except for the outline).

Bug: 444193399
Test: atest NotificationTest NotificationActionTest + screenshots
Flag: android.app.api_notification_action_custom
Change-Id: I55cbe93c2590c055b2abec17b598cd1d3f1f3949