Add plate overlay to bridged notification icons
This change implements the UI for bridged notifications by adding a
plate overlay to the small icon.

When a notification is bridged from another device (e.g., a phone),
this change decorates the small icon with a bottom plate and a symbol
of the originating device type.

Change-Id: I2b3254e744d2afcbd861195a01448ea00b985bc8
Bug: 438827600
Flag: android.app.bridged_notifications