[flexiglass] Don't clear input on bouncer decompose
Previously, we explicitly reset the text input when the auth
bouncer was decomposed. This also reset the input when the orientation
of the device changed. Since the viewmodel will be deleted when the
bouncer gets hidden, this is not needed since this already does the
input reset.

Bug: 457935502
Flag: com.android.systemui.scene_container
Test: atest Pin-/Pattern-/PasswordBouncerViewModelTest
Test: manual changed orientation on bouncer, dismissed and reopened
bouncer from lockscreen and notifications

Change-Id: Ib8b670de18fa57aa87d0c91b772e265e0924e28b