f/b OWNERS.md: emphasize root directory more
People often add OWNERS files in many places
in frameworks/base. This makes it harder to include
them, because you need to specify a deeply nested
path to include them, and it also makes it hard to
see which files are lists of teams and which files
are just including them. Many core teams use
frameworks/base/<TEAM>_OWNERS so recommend this as the
standard pattern.

Bug: N/A
Test: N/A
Flag: EXEMPT BUGFIX
Change-Id: Ibafcaaf97331f3802c6c1694fb91b5e791c58948