[Fix] Support normal layer requests without desktop env
This fix addresses an issue where both `desktopTasksController` and
`userRepositories` may be null on devices that do not support a desktop
environment. The layer request handler must be capable of processing
requests under these null conditions, as it is obligated to approve the
callback.

On such devices, the pinned layer is typically not available, which
prevents the user from exiting the normal layer. Consequently, a request
for a normal layer should be unconditionally accepted as a no-op.

Bug: 399376001
Bug: 449681882
Flag: com.android.window.flags.enable_interactive_picture_in_picture
Test: atest WMShellUnitTests:NormalAppLayerControllerTest
Change-Id: I120eb73b367a9c2fcb2fb110811f7334bb128335