[flexiglass] Do not use weighted aboveFold for BesideUserSwitcherLayout.
CL ag/36967088 introduced weighted handling for the aboveFold content
within FoldAware to resolve password spacer issues in StandardLayout.
However, this weighting caused a regression in
BesideUserSwitcherLayout (b/413303536). This CL resolves the regression
by adding the useWeightedAboveFold parameter to FoldAware, allowing
StandardLayout and BesideUserSwitcherLayout to configure the weighting
as required.

Bug: 460145182
Bug: 458579500

Test: manually test on IKA VM - verify that the password layout on
BesideUserSwitcherLayout is aligned in the center vertically.
Flag: EXEMPT BUGFIX

Change-Id: Ifa92a633a5f8fc039cd335ba09066e493229f50c