Rename supportsSwitchToByUser to isUiSwitchableHumanUser
We rename and clarify UserInfo.supportsSwitchToByUser(), to explain its
purpose: to indicate whether the given user should show up in switchable
UI presentations. This currently basically means that it is an (enabled)
FULL user, but for modularity, is based on the other switchable methods
in UserInfo, and therefore can apply even if we change our criteria
under-the-hood in the future.

Test: N/A. Just a renaming
Fixes: 456803282
Flag: EXEMPT PURE_REFACTOR
Change-Id: Ie95f613d0a4f6bf78fdde5c5627d9189d9328009