Check if correct display is off when attempting to take a screenshot
A task's screenshot is not taken if a display is off. Previously, we
were checking if the default display is off, but, with projected display
where a device can be off while content is still displayed on an
external display, we need to check the status of the display the task
being screenshotted is on.

Bug: 452853165
Test: Attempt to screenshot tasks while phone display is off while
connected to CD
Flag: EXEMPT BUGFIX

Change-Id: I1e8d755c1c2d92d3acd02ba58496c686ae6fa724