Add more robust thread handling to the fake executor.
This enforces that the scheduler is run on a specific thread, generally
the thread that the unit test is running on. Also enforce that
Future.get should never be called from that thread, because that would
deadlock the test.

This does mean that if we want to test any code that uses get() that the
test will have to be careful to run said code on a separate thread. But
there's not really any way around that, we can't have the test taking
precise control over the flow of time and then have it call code which
is expecting time to move forward.

Bug: 461796483
Test: mmma, atest LskfResetTests
Flag: android.app.lskfreset.flags.enable_lskf_reset_manager
Change-Id: Iec6007ed0271c544b8fa6a3cf619ca06f9460793