Make to/from filling changes EFFECTS_LIFECYCLE
Visibility calculations in EnsureActivitiesVisibleHelper make it so that
opaque filling containers that fill their parent are treated as
occluding and make things behind it invisible, so in a hierarchy like:

(top)
- Root Task (desk)
-- Task A (freeform mode, fullscreen bounds, opaque)
-- Task B (freeform mode, non-fullscreen bounds, opaque)
- Task C (fullscreen mode, opaque)

it should be expected that Task B and Task C are invisible. However,
changing bounds using WCT#setBounds did not actually trigger a
visibility recalculation, so changing Task A (non-fullscreen
bounds) to fullscreen bounds did not actually make B and C invisible.

If the visibility was recalculated for some other reason (such as A
being reorderd during a recents transition), then it would get to the
expected visibility of B and C, but then resizing Task A again with a
WCT would also fail to recalculate visibility and leave B and C
invisible, thus making the screen black behind non-fulscreen Task A.

This change checks if there was a bounds change in a WCT to/from
filling, and marks the transaction as TRANSIT_EFFECTS_LIFECYCLE to
ensure visibility is recalculated.

Flag: com.android.window.flags.make_filling_bounds_change_effect_lifecycle
Bug: 452347446
Test: using an immersive app in desktop (Clash of Clans) make it use
fullscreen bounds by clicking the "immersive" resize option in the
maximize menu - verify with dumpsys that B (gmail) and C (desktop
wallpaper activity) are invisible. Then swipe up to overview and
reselect Clash of Clans - verify dumpsys again. Finally resize it back
to original non-fullscreen bounds using the restore button - verify
gmail and desktop wallpaper and visible and not black.

Change-Id: I7019cfe903f7dbf52a4b617af0d79fd68bf5e335