Support AppTask.moveTaskTo API for pinned tasks
Pinned tasks can only be resized; they cannot be moved. When
AppTask.moveTaskTo is called for a pinned task, only the size (width and
height) of the requested bounds is applied, while the position is
ignored.

To meet security requirements, an application cannot control a pinned
task's position by resizing it. To enforce this, any subsequent resizing
is always done against the initial bounds (the ones before the first
resize). Therefore, this disallows any further manipulation because the
final bounds are determined as if only one resize has occurred (the last
one). The intermediate bounds are ignored. The initial position state
resets when the user moves the window. The window strives to move only
one edge per axis during resizing, and the edge moves outwards (it
expands towards the display edge, or contracts away from the display's
center). The only exception is when the window would end up off-screen;
in this case, its position is adjusted to keep the window fully
on-screen.

One-pager: go/ipip-resize
Bug: 399376001
Bug: 449119484
Flag: com.android.window.flags.enable_interactive_picture_in_picture
Test: atest WMShellUnitTests:PinnedLayerPresentationControllerTests
Change-Id: I7318fb18a5839d441321d3a2b3d5be1639754e4b