[flexiglass] Rewrite LSTI to support all possible Idle/Transition combinations.
The previous logic would ignore any Idle that came in after a) another Idle, b) a Transition that we didn't start a KTF transition for. With snapToScene (added after LSTI was originally written), this is not a guarantee, and CLs such as ag/36767795 exposed this issue.

There were also issues with Transitions and Idles coming in during overlay transitions, which could leave KTF stuck.

Additionally, add ownerReason logging to the KTF transitions so that if we receive bugreports with incorrect KTF state, we can figure out why that happened.

Bug: 458502412
Fixes: 459767733
Flag: com.android.systemui.scene_container
Test: atest LockscreenSceneTransitionInteractorTest
Test: atest SceneInteractorTest
Test: press power button while on Bouncer (should go to AOD, not lockscreen)
Test: press power button while swiping up to Bouncer (should go to AOD)
Test: press power button while swiping down from Bouncer (should still go to AOD)
Test: ...while swiping down on shade from gone
Test: ...while swiping down on shade from LS
Test: ...while swiping up on shade in above cases
Test: ...while unlocking from LS -> Gone
Test: ...while unlocking from Bouncer -> Gone
Test: ...while unlocking from Lockscreen -> Shade -> Shade+Bouncer -> Gone
Test: press power again to cancel Gone -> Lockscreen (should go AOD -> LS)
Test: with a local CL that snaps to Lockscreen from a transition from Shade -> Gone and vv.
Test: with a local CL that snaps to Lockscreen+Bouncer from Lockscreen
Test: with a local CL that snaps to Lockscreen from Lockscreen+Bouncer
Change-Id: Ibfbb57d32e7fbaa455c8e1ce4f4ab521983131a6