Reland "Intercept back navigation on Desk root task"
Set wct#setInterceptBackPressedOnTaskRoot() to true on the Desk when it
is created. This ensures that when a back navigation event occurs on a
desktop app, Core delegates control to Shell instead of triggering
moveTaskToBack() on the task.

This is part of the effort to stop desktop predictive-back-enabled apps
from being minimized when back navigation happens.
See go/desktop-back-nav-lld for detailed design.

Flag: com.android.window.flags.enable_back_navigation_desktop_app_no_minimize
Bug: 456302602
Test: Confirm back navigation on desktop apps not cause minimization
when flag is on

Change-Id: Iaaedc98dc71252aded17a7f7a9536c69518e6303