Prevent pinning shortcuts/widgets for app-locked packages.
The `isRequestPinItemSupported` and `isRequestPinAppWidgetSupported` methods now receive the calling package name. This package name is used to check if the calling application is currently under app lock, and if so, the request to pin a shortcut or app widget is denied.
Test for this change will be provided once migration of BaseShortcutManagerTests to junit4 is done.

Flag: android.content.pm.app_lock_shortcut_removal
Bug: 447568094
Test: na
Change-Id: I3c38886d1a36f252a2c6993d5f12fbb15edc75d4