a11y: Remove flag enable_magnify_magnification_key_gesture_dialog
This CL will remove the bug fix flag. Before this change, when this flag
is enabled, we will show `MagnifyMagnificationDialogDelegate`, which
will turn on Magnification when the dialog is poped up and also zoom in
full-screen mode. While it's disabled, we will show
`DefaultDialogDelegate`, which is only for show a dialog without turning
on Magnification feature and zoom-in full screen.

Bug: 449611324
Flag: EXEMPT FLAG_REMOVAL
Test: atest AccessibilityShortcutsRepositoryTest,
KeyGestureDialogStartableTest

Change-Id: Ibdbabb0b091abe500259c77af20ec601317cd93e