[flexiglass] Switch to occluded when awaking with an occluding app
Until now, we always stayed on lockscreen when waking up shortly
after going to sleep with an occluding app active. This led to
issues where keyguard elements were shown on top of the
occluding app (e.g. camera). This change fixes it by switching
to Scenes.Occluded when awaking with an occluding scene on top.

Bug: 460009776
Test: Manual awaking akita after going to sleep on secure camera
and calculator
Test: atest SceneContainerStartableTest
Flag: com.android.systemui.scene_container

Change-Id: Iab53685b8f50b436cb56fd92e931935d6662723f