Ensure ImeWindowToken is not visible by default
This modified the ImeWindowToken to not be visible by default, unless it
gets a child window that is also visible. This visibility would flow up
to the ImeContainer, which we check in a few cases The container being
visible by default led to some unintended behaviour changes.

This only modifies the client visibility of the ImeWindowToken, and
checks whether it is client visible as an additional requirement for
becoming visible and visibleRequested. This also adds a
shouldCheckTokenClientVisible for isVisible, similar to the existing
shouldCheckTokenVisibleRequested for isVisibleRequested, to make sure
the IME window itself cannot become visible if its window token is not
client visible.

This also extracts shouldCheckTokenClientVisible and
shouldCheckTokenVisibleRequested out of WindowState and into methods on
WindowToken that can be overriden by specific implementations.

Flag: android.view.inputmethod.warm_work_profile_ime
Bug: 293644773
Bug: 461896800
Bug: 461891273
Bug: 461890481
Bug: 461890462
Bug: 461881663
Bug: 461818096
Bug: 461783747
Test: atest MultiUserMockImeTest
  ImeWindowTokenTests
  BubbleNotifPanelOverExpandedBubblesSwipeUpClosesNotifPanel#testNotifPanelOverExpandedBubbles_swipeUpClosesNotifPanel
  AccessibilityShortcutChooserActivityTest#selectTestService_notPermittedByAdmin_blockedEvenIfNoWarningRequired
  AccessibilityShortcutChooserActivityTest#clickServiceTarget_notPermittedByAdmin_sendRestrictedDialogIntent
  AccessibilityShortcutChooserActivityTest#popEditShortcutMenuList_oneHandedModeDisabled_shouldNotBeInListView
  AccessibilityShortcutChooserActivityTest#popEditShortcutMenuList_oneHandedModeEnabled_shouldBeInListView
  AccessibilityShortcutChooserActivityTest#selectTestService_permissionDialog_uninstall_callsUninstaller_rowRemoved
  UpdateSelfViaIntentTest#noLaunchGrantPermission_update_success
  SimpleSaveActivityTest#testTapLink_tapBack_thenStartOverByTouchOutsideAndManualRequest
  NotificationTest#testClearAllNotification
Change-Id: I22e804e2e26d91014b4a141c30690264e9e6cff0