Record the number of times system_server has started since last reboot in the critical event log
This allows downstream tools to tell whether a system server start was
due to a device reboot or not. This allows for proper resolution of
which events in the critical event log are from the current device boot,
which is essential for timestamp resolution (critical event log uses
uptime). All this is helpful for ANR analysis.

Bug: 445910085
Flag: EXEMPT BUGFIX
Change-Id: Ie03159ed89075fe003b1ea40be8ebcfd8f02d7ff