Implement animation of media seekbar from squiggly to flat
This CL adds a transition ratio. It draws two tracks translating to the
left by an offset. The inactive track is the background track that
appears on the whole width of seekbar, while the active track is being
clipped on the thumb position to show the actual progress of media
track. When drawing the path, the transition ratio changes from 1f
to 0f as it the track passes the thumb position until it reaches 0f
after a wavelength distance. When transition ratio is zero, there is no
change in the y-axis and a straight line should be drawn through
x-axis.

Flag: com.android.systemui.media_controls_in_compose
Flag: com.android.systemui.scene_container
Fix: 461759426
Test: Checked UI and compared transition to the view version.
Change-Id: Ic8c3f502694a3e6b39dc61d03b1171ee8e7b1072