Add a prefix to the MediaRouter2 logging
A MediaRouter2 instance could be a local router or a proxy router.
The local router is a singleton but there could be multiple proxy
router instances. It is hard to debug the MediaRouter2 without
distinguishing the proxy router instances from the local router
instance. Thus, we would like to add a prefix to the MediaRouter2
logging. The prefix is "[local]" for the local router instance and
the prefix is "[proxy-clientPackageName]" for a proxy router
instance.

Flag: EXEMPT PURE_REFACTOR
Bug: b/330300516
Test: presubmit and manually tested on a phone
Change-Id: Iea74666a287984fa5174cb13999e4207b1492f21