Show opening windows during return animations.
By default, Shell sets the alpha of opening windows to 0, as the
handler itself should make them visible during the animation. Since
we're animating the closing window, we need to manually set the alpha
of the opening ones.

Note that return animations will only be enabled with the refactor,
so we don't need to do this in the old codepath.

Bug: 202516970
Flag: com.android.systemui.animation_library_shell_migration
Test: atest ActivityTransitionAnimator
Change-Id: Ie597dcf70eae07f82a981d8a99df00563bd1f247