Revert "Remove unused GesturePointerEventDetector/Listener classes"
Currently, SystemGesturesPointerEventListener runs in system_server
process, but there is a request to run it in a SystemUI process that can
be overridden by vendors. GesturePointerEventLitener is implemented as
the migration destination for SystemGesturesPointerEventListener within
SystemUi. It is not yet in use, but is planned to be used in the future

This reverts commit 540a0d1463fc808008031b93137474bcf937ee59.

Reason for revert: The migration was half way. To complete the migration, we need these classes.

Bug: b:460319287
Test: m
Flag: com.android.window.flags.enable_transient_gesture_in_system_ui

Change-Id: Idb71a33a9caf3da238bd6799c3013ced4d09b22a