Make some communal state flows start sharing while subscribed
This change makes two flows in the CommunalInteractor start sharing only
while subscribed instead of eagerly, which blocks the calling thread
until the first value emits.

The state value of isCommunalBlurring is read in other classes so this
change also makes sure that the StateFlow is being collected once SysUI
boots, and log its state changes in a table which would help with
debugging blur issues.

This likely addresses an ANR issue in the class at initialiation time.

Test: atest CommunalInteractorTest
Test: atest CommunalLoggerStartableTest
Test: manually verified communal blurring works correctly
Fix: 457201219
Flag: EXEMPT BUGFIX
Change-Id: I9bf9966a603e3abbd5a86c070f934e957f689cfb