[flexiglass] Adds DelegatingTransition to STL lib
This is a Transition (the first one being for ShowOrHideOverlay) that
can link to another transition and run with in parallel. This should be
useful for (and is indeed being used in the next CL) have two STLs where
one drives the other and vice-versa so they remain in sync even if they
don't have the same set of overlays.

Bug: 457668950
Test: see next CL in this chain
Flag: com.android.systemui.scene_container

Change-Id: I3083618e9ccdacb349662f3cd8a13fe08607fc45