Allow CC sessions creation from virtual devices.
Currently CC requires the device to be unlocked but the check is
always done on the default device. Virtual devices are never
considered secure, so if the session request originates from a
virtual device, we should allow the session to be created.

This allows for CC to be used directly from virtual devices.

Adding a safeguard to make sure that the UID is seen on the reported
deviceId, to eliminate the case where an app is using an explicitly
created device context for a different device.

Test: atest & manual
Bug: 481223389
Flag: EXEMPT BUGFIX

Change-Id: I52a849043563acb2157daf85cc21aa8162d20e7d