AOD notif icons now movable
The NotifIconContainerViewBinder does not support multiple bindings for
the same view, in this case AOD, because of how StatusBarIconView
inflation is managed.

By making the AOD icons a MovableElement, there is only a single binding
even if the Element is "placed" in multiple scenes.

Flag: com.android.systemui.scene_container
Fixes: 464082934
Test: manual
Change-Id: If4f6e86a83bf56ffea3a2563f7f7709fcb62f352