Flush PersisterQueue on system shutdown and user stopping
Those are the last two chances where PersisterQueue is able to update
the state on disks, so we should flush pending items immediately.

AppCompatConfigurationPersister uses a different PersisterQueue
instance, but it doesn't seem to be user-specific so there is no need to
flush it on user stopping.

Bug: 459499545
Test: Latest states of resumed activities with persistable mode as
persistAcrossReboots are preserved after reboots.
Test: atest PersisterQueueTests
Test: Closing a guest doesn't trigger the broadcast receiver.
Test: atest WindowManagerServiceTests
Flag: EXEMPT BUGFIX

Change-Id: I703d7f07d58f29080d16b59d9da41c9ce2d813ae