App Window Screenshot: Extract app window discovery logic
We add a new class, |AppWindowInteractor|, to hold the logic to get the
list of visible app windows in z-order. This logic will serve as the
source-of-truth for active windows. We chose to make this class
stateless as the task list are short-lived, which will remain in the
PreCaptureViewModel. This CL also extracts the tests related to app
window discovery to its own file.

Bug: 462874010
Test: atest PreCaptureViewModelTest
Test: atest AppWindowInteractorTest
Flag: com.android.systemui.large_screen_screencapture
Change-Id: Ic43b1575e2757994ca43f391d78a5c3966eb1218