Temporarily treat HSU as Admin in select places
The Headless System User should not be an Admin. But currently taking
away its Admin status breaks a few things. So we temporarily maintain
those places in the previous status quo, by explicitly terating the
system user is an Admin.
The goal is to eventually figure out proper solutions to these
places, but for now, we just maintain the status quo.

Bug: 457997251
Bug: 461646697
Bug: 457559134
Bug: 455113492
Flag: android.multiuser.hsu_not_admin
Test: atest BugreportManagerTest#testRemoteBugreport
Test: atest CtsDevicePolicyTestCases:android.devicepolicy.cts.BugReportTest
Test: atest CtsPackageManagerHostTestCases:PrivilegedUpdateTests
Change-Id: Iab54fbc35db16c51bafb46d7e54b5937a315454a