Show selected IME subtype on status bar IME chip.
To get the selected subtype for the active user, add a method
`selectedInputMethodSubtype` to `InputMethodRepository` which observes
settings changes and fetches the selected subtype accordingly.

For now, just try to show the selected subtype id on the IME chip. In
follow-ups, we'll adjust this so that it may show either the subtype
icon, short label, or a fallback keyboard icon depending on what the IME
has specified for the subtype.

Bug: 443854916
Bug: 458557858
Flag: com.android.systemui.status_bar_ime_chip
Test: atest SystemUITests:ImeIndicatorChipInteractorTest &&
  atest SystemUITests:ImeIndicatorChipViewModelTest &&
  atest SystemUITests:InputMethodRepositoryTest
Change-Id: If794e78dd439985d346f5b50bfd843a665ada7b0