[Flexiglass] Remove unused state from NotificationPlaceholderViewModel
This ViewModel exposed the quickSettings shade content key to the composables, which could use it for the decision of picking the right placeholder instance during transitions. Since this decision is moved to NotificationPlaceholderStateStorage's content pickers, the ViewModel doesn't need to expose this state anymore.

Bug: 462706428
Test: atest NotificationsPlaceholderViewModelTest
Flag: com.android.systemui.scene_container
Change-Id: I3f6756e8c879b216691ee8e6bb3773b72f9fa2bc