Animate smartspace card visibility changes
Normally, when the smartspace view is detached the sesssion is also
terminated. This causes visibility changes to the view to be terminated
until it is re-attached. Since compose controls whether the view is
attached, we keep the signal alive by listening for smartspace targets
directly. This persistent signal is a much more reliable way of telling
whether smartspace should be visible or not.

Bug: 466101295
Test: Checked AOD media animation
Flag: com.android.systemui.scene_container
Change-Id: Ifd9f40ae046972c5cc31c131f7f63c799a1e0f5b