An attempt at adding support for timeouts to the fake executor.
This allows callers to also use Future.get but with a timeout and not
just one that blocks indefinitely. This requires adding a decent amount
of complexity to the executor in order to track and things blocked on
the future executing versus things blocked on either that or a timeout,
including a bunch of extra internal synchronization.

Bug: 461796483
Test: mmma, atest LskfResetTests
Flag: android.app.lskfreset.flags.enable_lskf_reset_manager
Change-Id: Ie2b7753a36a178a5ed9664d848b8c16207c4c352