[FS Req] Add WindowContainer#mFullscreenRequestAllowMode
For WMShell to predefine which containers are allowed to request a
fullscreen mode entry or exit, so that requests that WMShell will never
approve don't need to start a transition and roundtrip to WMShell to
handle it.

By default, containers are set to ALLOW_NONE expect tasks which are set
to ALLOW_INHERIT to inherit their parent container mode. If none of the
ancestors have ALLOW_ENTER or ALLOW_EXIT then the container is
considered to not be allowed to handle either request (ALLOW_NONE).

In desktop windowing, the desk containers are set to ALLOW_ENTER so that
any desktop task is allowed to request fullscreen. Once a task
successfully moves to fullscreen then it is set to ALLOW_EXIT by
ClientFullscreenRequestController.

Flag: com.android.window.flags.delegate_request_fullscreen_handling_to_shell
Bug: 296268915
Test: using ApiDemos, request to enter and exit fullscreen from desktop
- verify requests succeed

Change-Id: I5ad516e3a97008759ecace4eb057c9da0cdbaa7b