Support two V3 scheme signers targeting the same SDK version
The Android 26Q2 release is introducing support for the ML-DSA
signature algorithm to sign APKs. Since this is the first release
with this support, new ML-DSA signers must target the current
development platform which is the same SDK version as the previously
released platform; the previous rotated signer would also need to
target the same SDK version to ensure contiguous support between
signers. Previous code would see that there are two signers
targeting the same SDK version and fail since the V3.0 / V3.1
signature schemes only support a single signer, even though the
latest signer is only targeting the development platform. This
commit adds support for this by allowing two signers targeting the
same SDK version as long as one of the signers is targeting the
release and the other the development version. In this case, the
signer targeting the development release is used since that is the
intended version as specified by the development attribute in the
V3.1 signing block.

Bug: 458115629
Flag: android.security.apk_pqc_hybrid_signing
Test: atest PkgInstallSignatureVerificationTest
Change-Id: I2048a059c02f3b07d3fcdb6ec95ab8e4589f22fe