Apply self-kill recovery only to top fillsParent activity
Currently if multiple activities are stucked in a task, self-kill
recovery runs for all of them.

- If the top activity is a transient popup activity, it may be
 expected to finish on display move, so we shouldn't consider it
 to be self kill.
- The self-kill logic currently runs even for invisible activities.

With this change, the logic will be limited to the "main"
activity, which should be sufficient for all cases we want to fix.

Flag: com.android.window.flags.enable_auto_recovery_from_self_kill
Bug: 446998828
Test: AppCompatDisplayCompatTests
Change-Id: If62a185210a2a0535d758408582ac5a594ba6042