[FS Req] Splitscreen: Handle fullscreen enter and exit requests
Adds a SplitScreenFullscreenRequestHandler to handle enter and exit
fullscreen mode requests from the Activity#requestFullscreenMode API.
Also sets WCT#setFullscreenRequestAllowMode on the split root so that
tasks in split are allowed to request to enter fullscreen in the first
place.

Flag: com.android.window.flags.delegate_request_fullscreen_handling_to_shell
Bug: 296268915
Test: using ApiDemos, put it and another app into split, request
fullscreen mode and verify it moves to fullscreen and the other app is
dismissed, then request fullscreen exit and verify it and the other app
are restored to their previous state.

Change-Id: Icf4c2b7964a3240f7ad51d7f3e84f5b98cedcfe9