Fix flickers when relaunching top app to bubble
The activity task was reparented to the bubble-root-task while
relaunching it to bubble. Since the task was not collected in
the transaction, the task surface was hidden because the
bubble-root-task was not yet shown once the pending transaction
was applied. This could be fixed by collecting the tasks and
defer the changes until sync transaction is applied.

In addition, the bubble task was always set to translucent
while start animating. Now, the task should remain visible
if it is just a change (visible to visible).

Bug: 465312426
Test: launch top app into bubble
Test: BubbleTransitionsTest
Flag: com.android.window.flags.enable_bubble_root_task
Change-Id: I94d5c0bbabcf9167c902eb7a3a55f0d6451359ff