Skip the automation warning dialog if automation has ended.
If an intent is resolved while an automation session is running,
but sent after the session has ended, just proceed with the launch
of the target.

Also improve the automated app interception - the current handling
is brittle and racy. Since we're closing the device anyway, clear
the repository for that device as soon as the user chooses to stop
the task.

Bug: 464160337
Flag: android.companion.virtualdevice.flags.computer_control_access
Test: atest, manual
Change-Id: Ia48e91ea23d1e533dde4c7b1d7df38873b1462d8