[flexiglass] Fix issues with keyguard enabledness.
This started as an app pinning investigation (app pinning works by disabling keyguard while the app is pinned, then re-enabling it and sending a lockNow() command). There was a bug where waking to a pinned app from a screen timeout (before the lock timeout) would incorrectly return to an unlocked homescreen.

However, the actual issue is that WM re-enables keyguard and sends the lockNow command at the same time, and order is not guaranteed. Legacy KeyguardViewMediator handles enabledness synchronously, while lockNow is always tossed onto a Handler thread, so we'd always receive lockNow() second.

Flexi handles these both async, so they can come in any order. Sometimes, we'd lockNow (while disabled), then re-enable. The lockNow() had no effect while disabled, and re-enabling did not re-show keyguard since it was not showing when it was disabled.

This exposes a further underlying bug, which is that actually, if we receive a lockNow() command while keyguard is disabled, we should re-show keyguard the next time it's enabled, even if it was not showing when it was originally disabled. KeyguardViewMediator implements this functionality; I'm not sure how/if it's used in legacy code other than for this race condition.

Fix this by adapting the existing KeyguardEnabledInteractor to support flexiglass, and deleting the duplicative reshowKeyguard boolean in the repository.

Fixes: 465772347
Test: atest SceneContainerStartableTest
Test: pin an app, allow the screen to time out, then wake it up prior to the lock timeout, swipe up to unpin, we should go back to lockscreen (note that you must dismiss the 'app is pinned' popup, app is *not* pinned until that's dismissed despite what it says, in which case you would expect to go back to unlocked)
Flag: com.android.systemui.scene_container
Change-Id: Idd995a32a9e30987d4235a7330e3ff178b2d71fe