LockedAppActivity on an external display
This change adds a message on the LockedAppActivity informing the user
if they need to authenticate on their other device. This will happen
when the user views the LockedAppActivity on an external display.

This change also resets and updates view on a configuration change.

Bug: 451909725
Flag: android.security.app_lock_apis
Test: manual - observed the message on the external display, but not on
  the main display
Test: atest LockedAppActivityTest
Change-Id: Idcc8a920bd37029c8bb8c6b35698693eccd4bd81