PrintSpooler: Reflect printer setup in UI
When a PrinterInfo has a setup Intent attached, print jobs cannot be
queued to that printer.  This change updates the UI to show that
printing is disabled in this case.  It turns the Print button into a
Setup button that triggers the attached Intent.

Bug: 460112550
Test: atest CtsPrintTestCases
Flag: android.print.flags.enable_setup_activity
Change-Id: If0110bb83b986b1204b580534bbecc8a87298a7c