[flexiglass] Don't immediately finish pulse when it's cancelled
When a user taps to expand a notification on AOD while pulsing,
the following sequence of events occur:

1. SysUI requests to wake the device & pulse cancelled
2. SysUI receives signal that device is waking up &
  finishes dozing & pulse finishes
4. Notification expands

Previously, we'd finish the pulse when Doze received the cancellation.
Now, we allow the pulse to finish when doze is finished. This is
to avoid a race condition in isShadeTouchable where expanding
a notification while pulsing would result in isShadeTouchable
briefly returning false becaues isPulsing was set to false
before the device woke up.

Test: atest DozeScrimControllerTest
Test: Expand a pulsing notification in AOD; observe
the device correctly wakes device and expands notification
Fixes: 454355544
Flag: com.android.systemui.scene_container

Change-Id: I2ce94380373e8c53ac0ee4c39eefe530e9d76758