[Cue Bar] Enable Dismissal Group IDs
Enables Dismissal Group IDs to prevent showing Cue Bar under the same
context.

Previously was guarded by a flag for which there exists a possiblity
that it will not be enabled by 26Q1.

This bug is evident before the introduction of the new feature (b/427284798) and becomes more evident if the feature's counterpart is enabled in AiAi go/cue-bar-on-ime-zero-state as there are deliberate multiple Cue Bar postings for the same context.

Bug: 427284798
Flag: EXEMPT BUGFIX
Test: atest frameworks/base/packages/SystemUI/multivalentTests/src/com/android/systemui/ambientcue/domain/interactor/AmbientCueInteractorTest.kt
Change-Id: I0c88e805b3bbcb1b4ab0822993422dc3da5d84e8