PrintSpooler: Update button text sooner
When switching from the PDF printer to a service-provided printer, the
button text changes from Save to Print.  However, because the button is
updated after the capabilities are loaded, the text can continue to show
the wrong string for a few seconds.  This is particularly noticeable on
the first use, but can still be observed later if a particular printer
takes a long time to fetch capabilities after previously selecting the
PDF printer.

Instead, update the button content right away before updating the rest
of the dialog.  This doesn't change functionality because the button is
disable or hidden until the capabilities are loaded, but it ensures that
the correct text or icon is shown in the meantime.

Bug: 460102645
Test: Change destination dropdown and monitor button behavior
Flag: com.android.printspooler.flags.updated_button_layout
Change-Id: Iba8d1e1d3e041a74ab5858e883e13f935663c623