[2/N] TaskPositioner refactor: Implement ResizeTaskPositioner
This is the orchestrator class. ResizeTaskPositioner is a new
implementation of the original TaskPositioner interface. It acts as a
controller that delegates the actual work to either a TaskMover or a
TaskResizer implementation based on the user's input (ctrlType). This is
where the core logic for handling drag start, move, and end events
resides.

See upcoming implementation at go/taskpositioner-refactor

Flag: com.android.window.flags.enable_task_positioner_refactor
Bug: 465979009
Test: atest ResizeTaskPositionerTest
Change-Id: Iaaef1c2767ee80e4e47a98b90f278b31dc0eb753