Avoid lockNow requests going missing
A race condition in DeviceUnlockedInteractor upon user switch can cause
a lockNow() request to be ignored if it comes in before the
authenticationMethod change to a secure authMethod has been processed.

Changing to a secure auth method adds the corresponding receiver on the
the lockNowRequests channel. Without this receiver, any trySend() call
from lockNow() would just silently fail as this channel is currently a
rendevouz channel. Failure here means that the lock request won't be
processed.

Making the channel for lockNowRequests conflated (i.e. bufferSize=1 with
BufferOverflow.DROP_OLDEST) fixes this as the last value would then be
processed when the lock request handlers are added when the
authenticationMethod change is processed.

Fix: 457867010
Test: atest DeviceUnlockedInteractorTest
Flag: com.android.systemui.scene_container
Change-Id: Ibe0185f468772e108f520874bf0b332f8a1e59ba