Add transient System UI dialogs to enable the overview gesture when showing a SystemUIDialog
The problems is that by default overview gesture is set to dismiss SystemUIDialog without opening the overview. This creates a poor UX with the dialogs that pass through touches and dismiss them selfs when this happens. For example, the Volume dialog (this one doesn't now inherit the SystemUIDialog for this exact reason) and the new Screen Recording Toolbar dialog.

Flag: com.android.systemui.new_screen_record_toolbar
Fixes: 463948190
Test: atest SystemUIDialogTest
Test: InputConsumerUtilsTest in presubmits
Test: Check that overview gesture dismisses the Screen Recording toolbar while performing their original motion
Test: Check that back gesture doesn't happen for the Screen Recording toolbar and is instead passed to whatever is behind it. Toolbar dismisses itself based on the interaction that happens out side of its bounds
Test: Check that other System UI dualogs behave as they used to
Change-Id: Ib0387886f19966340098983ea2dfb616f45b96f2