Disable MemoryLimiter tests in unsupported environments
The MemoryLimiter unit tests must be run with 'adb root' enabled.
This change adds a test at the front of each unit test that checks if
the environment is sound.  This also disabled MemoryLimiter unit tests
under Ravenwood.

AndroidTest.xml is not changed to enable 'adb root' because the
MemoryLimiter tests are part of a much larger suite of tests, and the
AndroidTest.xml file applies to all the tests.  The plan is to move
the MemoryLimiter tests into their own test "block", in a future CL.

Flag: com.android.server.am.memory_limiter_enable
Bug: 461546255
Test: atest
 * MemoryLimiterTest
Change-Id: I78d5b13411a6fdad0674798206b1fa6c41f56896