Move navigation bar window insets to BouncerOverlay
With containerized bouncer, this inset would currently be applied to the container instead of the outer wrapper. This CL moves the navigation bar window insets from BouncerContent to BouncerOverlay to fix that, but only when containerized. When not containerized, the navigation bar window insets are still applied to BouncerContent. Otherwise, the bouncer background would also be inset.

Fix: 469363430
Bug: 459987781
Test: manual
Flag: com.android.systemui.scene_container
Change-Id: Ibe1f52111ca30322aa4a6de537e8c2f416ead161