Fix transition when device canWakeDirectlyToGone
Fix the transition after a power button press when the device
canWakeDirectlyToGone. The target scene should be Gone, even if the
device is currently transitioning to the lockscreen.

This CL fixes integration tests like the ones in the associated bug, in
which the removed condition prevented the device from transitioning back
to the Gone scene, leaving the device in a pseudo-AOD scene.

Bug: 467015856
Test: atest PlatformScenarioTests:InitialSceneSingleShade
Flag: EXEMPT BUGFIX
Change-Id: I753cb07209fdce27ecff96fa2ef71a033642d138