Fixes relaunching bubbled app to Desktop
While relaunching a bubbled app into Desktop mode, the task was
reparented to desktop root-task, but the task bounds was remained
and displayed like a bubble. That's because the transition was
handled by the TaskViewTransitions and the task bounds was updated
as a bubble when the transition finished. This is fixed by avoid
handling it in TaskViewTransitions if the task has a parent Task.

In addition, this CL also clean-up the bubble data when the task
is relaunched to Desktop.

Bug: 447308997
Test: relaunch bubbled app in desktop
Test: wm presubmit
Flag: com.android.window.flags.enable_bubble_root_task

Change-Id: Ibb7dcfeb62576825bc49a63c4d745094d1ff6013