Cache the volume when the Assistant mode is set
We want to remember the last volume during the
MODE_ASSISTANT_CONVERSATION so that we can re-apply it whenever the mode
is reset. When exiting the mode and therefore switching the stream that
drives the absolute volume, it is possible that the assistant stream
will be lowered based on the new driving stream. This is why it is
important to cache the volume during the assistant mode and reapply it
whenever the assistant mode is reentered and the assistant stream was
not touched outside of this mode.

Test: manual with modified Gemini && atest VolumeHelperTest
Bug: 451627532
Flag: com.android.media.audio.abs_volume_stream_always_max
Change-Id: I97ebd305da54667158602d4d70ef980ec87e669e