Revert^2 "Finish the embedded visualizer implementation."
This reverts commit 43d713ebf28cd64f4cef761a6533177644c488eb.

Reason for revert: I have made changes to the SysuiVisualizerService such that it can be instantiated on all platforms (before it could only be instantiated on phones, which is why it was failing for non-phone platforms). I have also temporarily disabled the service. Here is the change: ag/37541261. This should make sure that it no longer breaks CTS tests (the reason for the original revert).

Bug: 461276741
Test: atest PersonalContextTests
Flag: android.service.personalcontext.enable_personal_context_service

Change-Id: I6fce4c35c18e94ea9aa06d3c24f9f898270957ea