Include test API surfaces in the aggregate signature files
Previously, this would not create aggregated signature files for `test`
APIs like it does for `system` and other APIs. This change adds them
for non-updatable modules (as they are the only ones that support
`test` APIs).

It was necessary to remove the `android.txt` and `removed.txt` dist
specification from `test-api-stubs-docs-non-updatable` as that
conflicted with the ones added for
`frameworks-base-api-test-exportable-current.txt`. This makes it
consistent with other API surfaces.

Flag: EXEMPT HOST_MODE
Bug: 469982585
Test: m frameworks-base-api-test-checked-in-current.txt
      # Verified that it only depends on checked in signature files.
      m frameworks-base-api-test-exportable-current.txt
      # Verified that it did not include flagged APIs.
Change-Id: I059baa0620490fa8b2f3f8ce2cbb578c54d5a3c9