Add option to exclude STL from debugging
In ag/37517423 a new STL was added that overlays the scenecontainer 1:1.
Unfortunately this now blocks the view on our most important STL
(SceneContainer).

This CL adds the option to exclude an STL from debugging and also
excludes this BouncerTopSTL per default. In the future we could explore
how we could offset these labels such that they don't overlap.

Bug: 463679218
Test: Excluding BouncerTopSTL works and is set per default on first run
Test: Built with Soong and Gradle
Flag: EXEMPT DEBUG
Change-Id: Ic278eb3981f0b55c51632f949efcab13bf50d824