Draw shared QsTiles in Qs scene for last 5% of Shade -> Qs
In b/377909993 the problem is that the transition Shade -> Qs takes a
while to settle. Until it is fully settled to Idle(Qs) Shade is
responsible to compose the SharedElement. Because the elements in shade
are not part of a context of a paginated tile container, you can not
fling to the next page until the transition is fully settled. To fix
this there are these options:

1) Somehow make Shade Tiles understand vertical gestures and pass this
to the QS Tile equivalents. This seems hard and messy.
2) Make Qs instead of Shade compose shared tiles. This is not possible
because Shade has a higher Z order. The tiles disappear behind the
Shade.
3) This solution. Hand over responsibility of drawing dynamically based
on progress. Most of the transition is still handled by shade, where
Z order matters. For the last 5% handover to Qs where gesture and
overscroll matters.

We have to disable the alpha fade of the Qs tile container for the
first 5% otherwise there is a small jump in alpha when drawing changes
because the Qs scene tiles are children of the QuickSettingsContent
which is faded out. 5% is small enough to barely affect the animation
while it is large enough to enable flings smoothly. I was not able to
create a fling that was eaten.

In b/466364382 one of the problems is that the shared tiles are part
of a very different hierarchy of containers and therefore the
overscroll effect is out of sync when you see shared/unshared elements
next to each other. Again, the same solution of handing over at 5%
solves this issue as now Qs is fully responsible for all tiles
overscroll so they are synced up correctly.

Bug: 377909993
Bug: 466364382
Test: Manual: First fling is not eaten anymore
Test: See video in b/466364382. No more stutter of the overscroll.
Flag: EXEMPT Bugfix
Change-Id: I0ac15583e9c58626c54ecec82887e4fae8d23209