adb uninstall all uses Admin user
adb shell pm uninstall <app> defaults to uninstalling from all users. In
this case, the userId that gets passed to the server becomes irrelevant,
since we'll be uninstalling from all users anyway.

However, that userId is used for an Admin check, since only an Admin can
downgrade system apps.

Historically, the userId has been hardcoded as the SYSTEM user, which
has been fine since the SYSTEM has been an Admin. But going forward, the
SYSTEM user may not be an Admin, which would cause this command to fail.
So we instead pick an actual Admin user.

Bug: 457997251
Flag: android.multiuser.hsu_not_admin
Test: atest CtsPackageManagerHostTestCases:PrivilegedUpdateTests

Change-Id: Ieec68a1f52206df7ab50451a957f8cde663d97ef