[sb] override appTransitionFinished and check for pending transitions
Ultimately, there are a number of issues with this code, the first of
which being this `forced` boolean in appTransitionStarting really has
_nothing_ to do with "forcing" anything, but rather it distinguishes
calls to `appTransition*` methods originating either internal or
external to our own process.

Secondarily, there are a lot of booleans inside of
LightBarTransitionsController that are attempting to keep track of the
following (non-exhaustive) list of things:
  1. what the "dark intensity" of the icons is (how light to make them)
  2. The timing of animations between dark intensities
  3. The doze amount, for when that is relevant
  4. whether or not we are in an app transition
  5. whether or not we are awaiting a transition of icon tint

I managed to track down this particular bug to a potential timing issue
when going from screen off -> app.
`LightBarTransitionsController.appTransitionStarting` gets called
_first_ internally (i.e., `forced = true`) and then immediately over the
binder interace (i.e., `forced = false`). This means that
appTransitionPending is called after the "forced" appTransitionStarting
happens.

When forced is false, we simply return early from the callback and never
run the animation that was queued up from `setIconsDark`. This leaves us
stuck waiting for another "forced" call that never comes.

We had some existing code in `appTransitionCanceled` which seemed to do
exactly what I wanted (check for a pending transition, and run it if it
exists), and an unused `appTransitionFinished` callback.

*The fix*
I copied the code from `appTransitionCanceled` into a helper function
that is called either on cancel or finish. And that pretty much covers
this bug. (we should still refactor all of this code. It's not doing
what you expect half the time).

*The "test"*
By adding a check and a log for (mTransitionPending &&
mTintChangePending) in the `appTransitionFinished` callback, I was able
to reproduce this issue locally and see that the log always happened
when the issue reproduced. That's as far as I've gotten to testing it.

Test: LightBarTransitionsControllerTest
Bug: 457816397
Flag: EXEMPT BUGFIX
Change-Id: Ic0e84e521dcc640bb108981e01ce04599b3e2f72