ComputerControl: Only show touches during foregrounded automation
When automation is running fully in the background, meaning there are no
active mirror views, there is no need to show touches. Showing touches
will result in the extra work of compositing the animations of touch
spots, which we can eliminate for the backgrounded case.

Bug: 456220219
Test: Presubmit, manual with computer control test app
Flag: android.companion.virtualdevice.flags.computer_control_access
Change-Id: I259caa6a2d218675692e0424d1b2d99af9995cb5