Prepare for profile triggering
The MemoryLimiter will trigger a profile on over-limit events. This
requires changes to the code to carry the package name around.  Unit
tests verify that the the package name (which may be null) is handled
properly.

This commit does not actually trigger a profile.

This also enhances the debug output as reported by the shell command
'am memory-limiter status'.

Flag: com.android.server.am.memory_limiter_default_app_limits
Bug: 467691174
Test: atest
 * com.android.server.am.MemoryLimiterTest
 * FrameworksServicesTests:com.android.server.am
 * FrameworksCoreTests:ActivityTests
 * FrameworksMockingServicesTests:com.android.server.am
Change-Id: I58aea3e407aa5a4178284eeb74db8dccc547da55