Augment/move ChangeBuilder and clean-up TransitionInfoBuilder
Some upcoming changes will require all changes to include
a WindowContainerToken. To prepare for this, move the
existing ChangeBuilder to the same utils location as
TransitionInfoBuilder, update ChangeBuilder to handle more
general usage and then use it to clean-up the
TransitionInfoBuilder a bit.

Also, made a number of other adjustments:
- removed confusing asNoOp in favor of making it like the
  real impl (add root-leash(s)) based on added changes.
- removed usages of nullable parameters where possible
- Inserted ChangeBuilder usage into TransitionInfoBuilder
- Replaced some of the addChange overloads which were
  only used a couple times with ChangeBuilder at callsites
- Added functionality to TestShellExecutor to work like
  the real ShellExecutor (where execute() is required to
  run synchronously on same thread).

Bug: 469422564
Test: existing tests
Flag: EXEMPT PURE_REFACTOR
Change-Id: I42b1afca2d4223ccf687aafd1867d7270ad0d05d