Don't relaunch the activity during drag-resizing if it is not drawn
Before this CL, when the user keeps changing the task size during
drag-resizing, the activity might be relaunched constantly without
drawing any frame.

This CL postpones any relaunch if the previous one is not finished or
the main window of the new activity is not drawn during drag-resizing.
In this way, the user will be able to see the layout change.

Bug: 429960813
Flag: com.android.window.flags.improve_fluid_resizing_performance
Test: perfetto trace
Change-Id: Ib5ce8a90630d7fd8ccd72630d6507c93fd13e199