Refresh SystemUIDialog background on uiMode change
This starts listening to uiMode changes between light and dark mode in
SystemUIDialog, and optionally refreshes the dialog's background with
the new theme color.

The background is normally coming from the Theme, and is set on the
DecorView. However, SystemUIDialog can use DialogTransitionAnimator for
animation handling, and that will set the backround on a different view,
and remove it from the DecorView.

This finds the right view that has the background, listens to the uiMode
changing, resolves the backround drawable and re-applies it.

This change only targets SystemUIDialogs with Compose content, as
Composables handle the change between light and dark mode directly,
while Views require manual handling.

Flag: com.android.systemui.dialog_background_refresh
Test: show the Modes dialog / Flashlight dialog / QS Reset dialog;
  adb shell cmd uimode night no
  adb shell cmd uimode night yes
Bug: 474600479
Change-Id: I4c3777394797abec32b43c858e13fee502270589