Introduce WindowAnimation to track transition animations
This change introduces a WindowAnimation class to track the animation of a single window during a shell transition. This allows the state of an animation (such as current bounds and scale) to be queried, which is a prerequisite for allowing one animation handler to seamlessly hand off control to another.

Test: atest DefaultTransitionHandlerTest
Bug: 477175273
Flag: EXEMPT PURE_REFACTOR
Change-Id: Iad63759b00689ed003f404a590b587848e707eaf