Set a min scale value for the NavBarPill animation.
Before the change, sometimes, the Row can report a event like:

```
12-30 14:23:57.335  2636  2636 I NavBarPill - Column - Row: Release, Offset(Infinity, Infinity)
```

And after this event, this Row won't receive any further touch / click signals, which will make the cue bar unresponsive.

This can be triggered stably by pulling down the notification shade. 

After the fix, the Row will not report this `Offset(Infinity, Infinity)`.

Note, the issue is not reproducible in the BP4A build. This might be an issue introduced by the recent compose drop.

Bug: 470164522
Change-Id: Ie06dc318d5bcb5464fcb21137befa443e4480a68
Flag: EXEMPT BUGFIX
Test: local build and run